@@ -1,81 +1,70 @@
 <?php
 
 /**
- * Class Editorial_Metadata
+ * Class EditorialMetadata
  * Editorial Metadata for VIP Workflow
  */
 namespace VIPWorkflow\Modules;
 
-require_once __DIR__ . '/rest/editorial-metadata.php';
+require_once __DIR__ . '/rest/editorial-metadata-endpoint.php';
 
+use VIPWorkflow\Modules\EditorialMetadata\REST\EditorialMetadataEndpoint;
+use VIPWorkflow\Modules\Shared\PHP\HelperUtilities;
+use VIPWorkflow\Modules\Shared\PHP\InstallUtilities;
 use VIPWorkflow\VIP_Workflow;
-use VIPWorkflow\Modules\Shared\PHP\Module;
-use VIPWorkflow\Modules\EditorialMetadata\REST\EditEditorialMetadata;
-
 use WP_Error;
+use WP_Term;
 
-class Editorial_Metadata extends Module {
+class EditorialMetadata {
 
-	// ToDo: Add date, and user as supported metadata types
 	const SUPPORTED_METADATA_TYPES = [
 		'checkbox',
 		'text',
+		'date',
 	];
-	const METADATA_TAXONOMY = 'vw_editorial_meta';
-	const METADATA_POSTMETA_KEY = 'vw_editorial_meta';
+	const METADATA_TAXONOMY        = 'vw_editorial_meta';
+	const SETTINGS_SLUG            = 'vw-editorial-metadata';
+	const METADATA_TYPE_KEY        = 'type';
+	const METADATA_POSTMETA_KEY    = 'postmeta_key';
 
-	public $module;
+	private static $editorial_metadata_terms_cache = [];
 
-	private $editorial_metadata_terms_cache = [];
-
-	public function __construct() {
-		// Register the module with VIP Workflow
-		$this->module_url = $this->get_module_url( __FILE__ );
-		$args             = [
-			'title'                 => __( 'Editorial Metadata', 'vip-workflow' ),
-			'short_description'     => __( 'Configure your editorial metadata for each post.', 'vip-workflow' ),
-			'extended_description'  => __( 'Editorial metadata can be used to enhance the details logged with each post. It is completely customizable for your workflow, using text and checkbox fields as supported field types.', 'vip-workflow' ),
-			'module_url'            => $this->module_url,
-			'slug'                  => 'editorial-metadata',
-			'configure_page_cb'    => 'print_configure_view',
-		];
-		$this->module     = VIP_Workflow::instance()->register_module( 'editorial_metadata', $args );
-	}
-
-	public function init() {
-		// Register the taxonomy we use for Editorial Metadata with WordPress core
-		$this->register_taxonomy();
-
-		EditEditorialMetadata::init();
-
-		// Load CSS and JS resources that we probably need in the admin page
-		add_action( 'admin_enqueue_scripts', [ $this, 'action_admin_enqueue_scripts' ] );
-
-		// Load block editor JS
-		add_action( 'enqueue_block_editor_assets', [ $this, 'load_scripts_for_block_editor' ] );
-
-		// Load block editor CSS
-		add_action( 'enqueue_block_editor_assets', [ $this, 'load_styles_for_block_editor' ] );
+	public static function init(): void {
+		// Register the taxonomy we use for Editorial Metadata with WordPress core, and ensure its registered before custom status
+		add_action( 'init', [ __CLASS__, 'register_editorial_metadata_taxonomy' ] );
 
 		// Register the post meta for each editorial metadata term
-		add_action( 'init', [ $this, 'register_editorial_metadata_terms_as_post_meta' ] );
+		add_action( 'init', [ __CLASS__, 'register_editorial_metadata_terms_as_post_meta' ] );
+
+		// Setup editorial metadata on first install
+		add_action( 'init', [ __CLASS__, 'setup_install' ] );
+
+		// Add menu sidebar item
+		add_action( 'admin_menu', [ __CLASS__, 'add_admin_menu' ] );
+
+		// Load CSS and JS resources for the admin page
+		add_action( 'admin_enqueue_scripts', [ __CLASS__, 'action_admin_enqueue_scripts' ] );
+
+		// Load block editor JS
+		add_action( 'enqueue_block_editor_assets', [ __CLASS__, 'load_scripts_for_block_editor' ] );
+
+		// Load block editor CSS
+		add_action( 'enqueue_block_editor_assets', [ __CLASS__, 'load_styles_for_block_editor' ] );
 	}
 
 	/**
 	 * Register the post meta for each editorial metadata term
+	 *
+	 * @access private
 	 */
-	public function register_editorial_metadata_terms_as_post_meta() {
-		$editorial_metadata_terms = $this->get_editorial_metadata_terms();
+	public static function register_editorial_metadata_terms_as_post_meta(): void {
+		$editorial_metadata_terms = self::get_editorial_metadata_terms();
 
 		foreach ( $editorial_metadata_terms as $term ) {
-			$post_meta_key = $this->get_postmeta_key( $term );
-			$post_meta_args = $this->get_postmeta_args( $term );
+			$post_meta_key  = $term->meta[ self::METADATA_POSTMETA_KEY ];
+			$post_meta_args = self::get_postmeta_args( $term );
 
-			// ToDo: If a post type was supported, a post was opened and saved, and then the post type was removed from the supported list,
-			// the post meta will still exist for new posts of that post type for a short duration.
-			// It works fine otherwise. This is a limitation of the current implementation.
-			// Working around this is quite expensive, and is fine for now.
-			foreach ( $this->get_supported_post_types() as $post_type ) {
+			foreach ( HelperUtilities::get_supported_post_types() as $post_type ) {
 				register_post_meta( $post_type, $post_meta_key, $post_meta_args );
 			}
 		}
@@ -83,109 +72,148 @@ class Editorial_Metadata extends Module {
 
 	/**
 	 * Register the post metadata taxonomy
+	 *
+	 * @access private
 	 */
-	public function register_taxonomy() {
+	public static function register_editorial_metadata_taxonomy(): void {
 		// We need to make sure taxonomy is registered for all of the post types that support it
-		$supported_post_types = $this->get_supported_post_types();
+		$supported_post_types = HelperUtilities::get_supported_post_types();
 
 		register_taxonomy( self::METADATA_TAXONOMY, $supported_post_types,
-			array(
-				'public' => false,
-				'labels' => array(
-					'name' => _x( 'Editorial Metadata', 'taxonomy general name', 'vip-workflow' ),
+			[
+				'public'  => false,
+				'labels'  => [
+					'name'          => _x( 'Editorial Metadata', 'taxonomy general name', 'vip-workflow' ),
 					'singular_name' => _x( 'Editorial Metadata', 'taxonomy singular name', 'vip-workflow' ),
-					'search_items' => __( 'Search Editorial Metadata', 'vip-workflow' ),
+					'search_items'  => __( 'Search Editorial Metadata', 'vip-workflow' ),
 					'popular_items' => __( 'Popular Editorial Metadata', 'vip-workflow' ),
-					'all_items' => __( 'All Editorial Metadata', 'vip-workflow' ),
-					'edit_item' => __( 'Edit Editorial Metadata', 'vip-workflow' ),
-					'update_item' => __( 'Update Editorial Metadata', 'vip-workflow' ),
-					'add_new_item' => __( 'Add New Editorial Metadata', 'vip-workflow' ),
+					'all_items'     => __( 'All Editorial Metadata', 'vip-workflow' ),
+					'edit_item'     => __( 'Edit Editorial Metadata', 'vip-workflow' ),
+					'update_item'   => __( 'Update Editorial Metadata', 'vip-workflow' ),
+					'add_new_item'  => __( 'Add New Editorial Metadata', 'vip-workflow' ),
 					'new_item_name' => __( 'New Editorial Metadata', 'vip-workflow' ),
-				),
+				],
 				'rewrite' => false,
-			)
+			]
 		);
 	}
 
 	/**
 	 * Load default editorial metadata the first time the module is loaded
+	 *
+	 * @access private
 	 */
-	public function install() {
-		// ToDo: Review the default metadata fields we provide OOB
-		$default_terms = [
-			[
-				'name' => __( 'Assignment', 'vip-workflow' ),
-				'slug' => 'assignment',
-				'type' => 'text',
-				'description' => __( 'What the post needs to cover.', 'vip-workflow' ),
-			],
-			[
-				'name' => __( 'Needs Photo', 'vip-workflow' ),
-				'slug' => 'needs-photo',
-				'type' => 'checkbox',
-				'description' => __( 'Checked if this post needs a photo.', 'vip-workflow' ),
-			],
-			[
-				'name' => __( 'Word Count', 'vip-workflow' ),
-				'slug' => 'word-count',
-				'type' => 'text',
-				'description' => __( 'Required post length in words.', 'vip-workflow' ),
-			],
-		];
+	public static function setup_install(): void {
+		InstallUtilities::install_if_first_run( self::SETTINGS_SLUG, function () {
+			$default_terms = [
+				[
+					'name'        => __( 'Embargo Date', 'vip-workflow' ),
+					'slug'        => 'embargo-date',
+					'type'        => 'date',
+					'description' => __( 'The date a post can be published.', 'vip-workflow' ),
+				],
+				[
+					'name'        => __( 'Expiry Date', 'vip-workflow' ),
+					'slug'        => 'expiry-date',
+					'type'        => 'date',
+					'description' => __( 'The date the content of a post would be marked as expired.', 'vip-workflow' ),
+				],
+				[
+					'name'        => __( 'Legal Review', 'vip-workflow' ),
+					'slug'        => 'legal-review',
+					'type'        => 'checkbox',
+					'description' => __( 'This is to be checked once legal has reviewed this post.', 'vip-workflow' ),
+				],
+			];
 
-		// Load the metadata fields if the slugs don't conflict
-		foreach ( $default_terms as $term ) {
-			if ( ! term_exists( $term['slug'], self::METADATA_TAXONOMY ) ) {
-				$this->insert_editorial_metadata_term( $term );
+			// Load the metadata fields if the slugs don't conflict
+			foreach ( $default_terms as $term ) {
+				if ( ! term_exists( $term['slug'], self::METADATA_TAXONOMY ) ) {
+					self::insert_editorial_metadata_term( $term );
+				}
 			}
-		}
+		});
+	}
+
+	/**
+	 * Register admin sidebar menu
+	 *
+	 * @access private
+	 */
+	public static function add_admin_menu(): void {
+		$menu_title = __( 'Editorial Metadata', 'vip-workflow' );
+
+		add_submenu_page( CustomStatus::SETTINGS_SLUG, $menu_title, $menu_title, 'manage_options', self::SETTINGS_SLUG, [ __CLASS__, 'render_settings_view' ] );
+	}
+
+	/**
+	 * Print settings page for the Editorial Metadata module
+	 *
+	 * @access private
+	 */
+	public static function render_settings_view(): void {
+		include_once __DIR__ . '/views/manage-editorial-metadata.php';
 	}
 
 	/**
 	 * Enqueue resources that we need in the admin settings page
+	 *
+	 * @access private
 	 */
-	public function action_admin_enqueue_scripts() {
+	public static function action_admin_enqueue_scripts(): void {
 		// Load Javascript we need to use on the configuration views
-		if ( $this->is_whitelisted_settings_view() ) {
+		if ( HelperUtilities::is_settings_view_loaded( self::SETTINGS_SLUG ) ) {
 			$asset_file = include VIP_WORKFLOW_ROOT . '/dist/modules/editorial-metadata/editorial-metadata-configure.asset.php';
 			wp_enqueue_script( 'vip-workflow-editorial-metadata-configure', VIP_WORKFLOW_URL . 'dist/modules/editorial-metadata/editorial-metadata-configure.js', $asset_file['dependencies'], $asset_file['version'], true );
 			wp_enqueue_style( 'vip-workflow-editorial-metadata-styles', VIP_WORKFLOW_URL . 'dist/modules/editorial-metadata/editorial-metadata-configure.css', [ 'wp-components' ], $asset_file['version'] );
 
 			wp_localize_script( 'vip-workflow-editorial-metadata-configure', 'VW_EDITORIAL_METADATA_CONFIGURE', [
-				'supported_metadata_types' => self::SUPPORTED_METADATA_TYPES,
-				'editorial_metadata_terms' => $this->get_editorial_metadata_terms(),
-				'url_edit_editorial_metadata'    => EditEditorialMetadata::get_crud_url(),
+				'supported_metadata_types'    => self::SUPPORTED_METADATA_TYPES,
+				'editorial_metadata_terms'    => self::get_editorial_metadata_terms(),
+				'url_edit_editorial_metadata' => EditorialMetadataEndpoint::get_url(),
 			] );
 		}
 	}
 
-	public function load_scripts_for_block_editor() {
-		$asset_file = include VIP_WORKFLOW_ROOT . '/dist/modules/editorial-metadata/editorial-metadata-block.asset.php';
+	/**
+	 * Enqueue resources that we need in the admin settings page
+	 *
+	 * @access private
+	 */
+	public static function load_scripts_for_block_editor(): void {
+		$asset_file   = include VIP_WORKFLOW_ROOT . '/dist/modules/editorial-metadata/editorial-metadata-block.asset.php';
 		$dependencies = array_merge( $asset_file['dependencies'], [ 'vip-workflow-block-custom-status-script' ] );
 
 		wp_enqueue_script( 'vip-workflow-block-editorial-metadata-script', VIP_WORKFLOW_URL . 'dist/modules/editorial-metadata/editorial-metadata-block.js', $dependencies, $asset_file['version'], true );
 
 		wp_localize_script( 'vip-workflow-block-editorial-metadata-script', 'VW_EDITORIAL_METADATA', [
-			'editorial_metadata_terms' => $this->get_editorial_metadata_terms(),
+			'editorial_metadata_terms' => self::get_editorial_metadata_terms(),
 		] );
 	}
 
-	public function load_styles_for_block_editor() {
+	/**
+	 * Enqueue resources that we need in the block editor
+	 *
+	 * @access private
+	 */
+	public static function load_styles_for_block_editor(): void {
 		$asset_file = include VIP_WORKFLOW_ROOT . '/dist/modules/editorial-metadata/editorial-metadata-block.asset.php';
 
-		wp_enqueue_style( 'vip-workflow-manager-styles', VIP_WORKFLOW_URL . 'dist/modules/editorial-metadata/editorial-metadata-block.css', [ 'wp-components' ], $asset_file['version'] );
+		wp_enqueue_style( 'vip-workflow-editorial-metadata-styles', VIP_WORKFLOW_URL . 'dist/modules/editorial-metadata/editorial-metadata-block.css', [ 'wp-components' ], $asset_file['version'] );
 	}
 
+	// Public API
+
 	/**
-	 * Get all of the editorial metadata terms as objects and sort by position
+	 * Get all of the editorial metadata terms as objects
 	 *
 	 * @param array $filter_args Filter to specific arguments
 	 * @return array $ordered_terms The terms as they should be ordered
 	 */
-	public function get_editorial_metadata_terms() {
+	public static function get_editorial_metadata_terms(): array {
 		// Internal object cache for repeat requests
-		if ( ! empty( $this->editorial_metadata_terms_cache ) ) {
-			return $this->editorial_metadata_terms_cache;
+		if ( ! empty( self::$editorial_metadata_terms_cache ) ) {
+			return self::$editorial_metadata_terms_cache;
 		}
 
 		$terms = get_terms( [
@@ -198,124 +226,122 @@ class Editorial_Metadata extends Module {
 			$terms = [];
 		}
 
-		$ordered_terms = [];
-		$hold_to_end = [];
-
-		// Order the terms
-		foreach ( $terms as $key => $term ) {
-			// Unencode and set all of our psuedo term meta because we need the position if it exists
-			$unencoded_description = $this->get_unencoded_description( $term->description );
-			if ( is_array( $unencoded_description ) ) {
-				foreach ( $unencoded_description as $key => $value ) {
-					$term->$key = $value;
-				}
-			}
-
-			// We require the position key later on
-			if ( ! isset( $term->position ) ) {
-				$term->position = false;
-			}
-
-			// Set the post meta key for the term, this is not set when the term is first created due to a lack of term_id.
-			if ( ! isset( $term->meta_key ) ) {
-				$term->meta_key = $this->get_postmeta_key( $term );
-			}
-
-			// Only add the term to the ordered array if it has a set position and doesn't conflict with another key
-			// Otherwise, hold it for later
-			if ( $term->position && ! array_key_exists( $term->position, $ordered_terms ) ) {
-				$ordered_terms[ (int) $term->position ] = $term;
-			} else {
-				$hold_to_end[] = $term;
-			}
-		}
-		// Sort the items numerically by key
-		ksort( $ordered_terms, SORT_NUMERIC );
-		// Append all of the terms that didn't have an existing position
-		foreach ( $hold_to_end as $unpositioned_term ) {
-			$ordered_terms[] = $unpositioned_term;
-		}
-
-		$ordered_terms = array_values( $ordered_terms );
+		$terms = array_map( [ __CLASS__, 'add_metadata_to_term' ], $terms );
 
 		// Set the internal object cache
-		$this->editorial_metadata_terms_cache = $ordered_terms;
+		self::$editorial_metadata_terms_cache = $terms;
 
-		return $ordered_terms;
+		return $terms;
 	}
 
 	/**
 	 * Returns a term for single metadata field
 	 *
-	 * @param int|string $field The slug or ID for the metadata field term to return
+	 * @param string $field The field to search by
+	 * @param int|string $value The value to search for
 	 * @return WP_Term|false $term Term's object representation
 	 */
-	public function get_editorial_metadata_term_by( $field, $value ) {
+	public static function get_editorial_metadata_term_by( string $field, int|string $value ): WP_Term|false {
 		// We only support id, slug and name for lookup.
 		if ( ! in_array( $field, [ 'id', 'slug', 'name' ] ) ) {
 			return false;
 		}
 
+		$term = false;
+
 		if ( 'id' === $field ) {
-			$field = 'term_id';
+			$term = get_term( $value, self::METADATA_TAXONOMY );
+		} else {
+			$term = get_term_by( $field, $value, self::METADATA_TAXONOMY );
 		}
 
-		// ToDo: This is inefficient as we are fetching all the terms, and then finding the one that matches.
-		$terms = $this->get_editorial_metadata_terms();
-		$term = wp_filter_object_list( $terms, [ $field => $value ] );
+		if ( is_wp_error( $term ) || ! $term ) {
+			$term = false;
+		} else {
+			$term = self::add_metadata_to_term( $term );
+		}
 
-		$term = array_shift( $term );
+		return $term;
+	}
 
-		return null !== $term ? $term : false;
+	/**
+	 * Add all the metadata fields to a term
+	 *
+	 * @param WP_Term $term The term to add metadata to
+	 * @return WP_Term $term The term with metadata added
+	 */
+	public static function add_metadata_to_term( WP_Term $term ): WP_Term {
+		if ( ! isset( $term->taxonomy ) || self::METADATA_TAXONOMY !== $term->taxonomy ) {
+			return $term;
+		}
+
+		// if metadata is already set, don't overwrite it
+		if ( isset( $term->meta ) && isset( $term->meta[ self::METADATA_TYPE_KEY ] ) && isset( $term->meta[ self::METADATA_POSTMETA_KEY ] ) ) {
+			return $term;
+		}
+
+		$term_meta = [];
+		$term_meta[ self::METADATA_TYPE_KEY ] = get_term_meta( $term->term_id, self::METADATA_TYPE_KEY, true );
+		$term_meta[ self::METADATA_POSTMETA_KEY ] = get_term_meta( $term->term_id, self::METADATA_POSTMETA_KEY, true );
+
+		if ( '' === $term_meta[ self::METADATA_TYPE_KEY ] || '' === $term_meta[ self::METADATA_POSTMETA_KEY ] ) {
+			return $term;
+		}
+
+		$term->meta = $term_meta;
+
+		return $term;
 	}
 
 	/**
 	 * Insert a new editorial metadata term
-	 * @todo Handle conflicts with existing terms at that position (if relevant)
+	 *
+	 * @param array $args The arguments for the new term
+	 * @return WP_Term|WP_Error $term The new term or a WP_Error object if something disastrous happened
 	 */
-	public function insert_editorial_metadata_term( $args ) {
-		// Term is always added to the end of the list
-		$default_position = count( $this->get_editorial_metadata_terms() ) + 1;
-
-		$defaults = [
-			'position' => $default_position,
-			'name' => '',
-			'slug' => '',
-			'description' => '',
-			'type' => '',
+	public static function insert_editorial_metadata_term( array $args ): WP_Term|WP_Error {
+		$term_to_save = [
+			'slug'        => $args['slug'] ?? sanitize_title( $args['name'] ),
+			'description' => $args['description'] ?? '',
 		];
-		$args = array_merge( $defaults, $args );
 		$term_name = $args['name'];
-		unset( $args['name'] );
 
-		// We're encoding metadata that isn't supported by default in the term's description field
-		$args_to_encode = [
-			'description' => $args['description'],
-			'position' => $args['position'],
-			'type' => $args['type'],
-		];
+		$inserted_term = wp_insert_term( $term_name, self::METADATA_TAXONOMY, $term_to_save );
 
-		$encoded_description = $this->get_encoded_description( $args_to_encode );
-		$args['description'] = $encoded_description;
-
-		unset( $args['position'] );
-		unset( $args['type'] );
-
-		$inserted_term = wp_insert_term( $term_name, self::METADATA_TAXONOMY, $args );
-
-		// Reset the internal object cache
-		$this->editorial_metadata_terms_cache = [];
-
-		// Populate the inserted term with the new values, or else only the term_taxonomy_id and term_id are returned.
 		if ( is_wp_error( $inserted_term ) ) {
 			return $inserted_term;
-		} else {
-			// Update the term with the meta_key, as we use the term_id to generate it
-			$this->update_editorial_metadata_term( $inserted_term['term_id'] );
-			$inserted_term = $this->get_editorial_metadata_term_by( 'id', $inserted_term['term_id'] );
 		}
 
-		return $inserted_term;
+		// Reset the internal object cache
+		self::$editorial_metadata_terms_cache = [];
+
+		$term_id = $inserted_term['term_id'];
+
+		$metadata_type = $args['type'];
+		$metadata_postmeta_key  = self::get_postmeta_key( $metadata_type, $term_id );
+
+		$type_meta_result = add_term_meta( $term_id, self::METADATA_TYPE_KEY, $metadata_type );
+		if ( is_wp_error( $type_meta_result ) ) {
+			return $type_meta_result;
+		} else if ( ! $type_meta_result ) {
+			// If we can't save the type, we should delete the term
+			wp_delete_term( $term_id, self::METADATA_TAXONOMY );
+			return new WP_Error( 'invalid', __( 'Unable to create editorial metadata.', 'vip-workflow' ) );
+		}
+
+		$postmeta_meta_result = add_term_meta( $term_id, self::METADATA_POSTMETA_KEY, $metadata_postmeta_key );
+		if ( is_wp_error( $postmeta_meta_result ) ) {
+			return $postmeta_meta_result;
+		} else if ( ! $postmeta_meta_result ) {
+			// If we can't save the postmeta key, we should delete the term
+			delete_term_meta( $term_id, self::METADATA_TYPE_KEY );
+			wp_delete_term( $term_id, self::METADATA_TAXONOMY );
+			return new WP_Error( 'invalid', __( 'Unable to create editorial metadata.', 'vip-workflow' ) );
+		}
+
+		$term_result = self::get_editorial_metadata_term_by( 'id', $term_id );
+
+		return $term_result;
 	}
 
 	/**
@@ -323,57 +349,36 @@ class Editorial_Metadata extends Module {
 	 *
 	 * @param int $term_id The term's unique ID
 	 * @param array $args Any values that need to be updated for the term
-	 * @return object|WP_Error $updated_term The updated term or a WP_Error object if something disastrous happened
+	 * @return WP_Term|WP_Error $updated_term The updated WP_Term or a WP_Error object if something disastrous happened
 	*/
-	public function update_editorial_metadata_term( $term_id, $args = [] ) {
-		$old_term = $this->get_editorial_metadata_term_by( 'id', $term_id );
-		if ( ! $old_term ) {
-			return new WP_Error( 'invalid', __( "Editorial metadata term doesn't exist.", 'vip-workflow' ) );
+	public static function update_editorial_metadata_term( int $term_id, array $args = [] ): WP_Term|WP_Error {
+		$old_term = self::get_editorial_metadata_term_by( 'id', $term_id );
+		if ( is_wp_error( $old_term ) ) {
+			return $old_term;
+		} else if ( ! $old_term ) {
+			return new WP_Error( 'invalid', __( "Editorial metadata doesn't exist.", 'vip-workflow' ) );
 		}
 
 		// Reset the internal object cache
-		$this->editorial_metadata_terms_cache = [];
+		self::$editorial_metadata_terms_cache = [];
 
-		$new_args = [];
+		$term_fields_to_update = [
+			'name'    => isset( $args['name'] ) ? $args['name'] : $old_term->name,
+			'slug'    => isset( $args['slug'] ) ? $args['slug'] : $old_term->slug,
+			'description' => isset( $args['description'] ) ? $args['description'] : $old_term->description,
+		];
 
-		$old_args = array(
-			'position' => $old_term->position,
-			'name' => $old_term->name,
-			'slug' => $old_term->slug,
-			'description' => $old_term->description,
-			'type' => $old_term->type,
-			'meta_key' => isset( $old_term->meta_key ) ? $old_term->meta_key : $this->get_postmeta_key( $old_term ),
-		);
+		$updated_term = wp_update_term( $term_id, self::METADATA_TAXONOMY, $term_fields_to_update );
 
-		$new_args = array_merge( $old_args, $args );
-
-		// We're encoding metadata that isn't supported by default in the term's description field
-		$args_to_encode = array(
-			'description' => $new_args['description'],
-			'position' => $new_args['position'],
-			'type' => $new_args['type'],
-			'meta_key' => $new_args['meta_key'],
-		);
-		$encoded_description = $this->get_encoded_description( $args_to_encode );
-		$new_args['description'] = $encoded_description;
-
-		unset( $new_args['position'] );
-		unset( $new_args['type'] );
-		unset( $new_args['meta_key'] );
-
-		$updated_term = wp_update_term( $term_id, self::METADATA_TAXONOMY, $new_args );
-
-		// Reset the internal object cache
-		$this->editorial_metadata_terms_cache = [];
-
-		// Populate the updated term with the new values, or else only the term_taxonomy_id and term_id are returned.
 		if ( is_wp_error( $updated_term ) ) {
 			return $updated_term;
-		} else {
-			$updated_term = $this->get_editorial_metadata_term_by( 'id', $term_id );
 		}
 
-		return $updated_term;
+		// No need to update the metadata as type can't be changed, and so neither can the postmeta key.
+
+		$term_result = self::get_editorial_metadata_term_by( 'id', $term_id );
+
+		return $term_result;
 	}
 
 	/**
@@ -382,56 +387,57 @@ class Editorial_Metadata extends Module {
 	 * @param int $term_id The term we want deleted
 	 * @return bool $result Whether or not the term was deleted
 	 */
-	public function delete_editorial_metadata_term( $term_id ) {
-		$term = $this->get_editorial_metadata_term_by( 'id', $term_id );
-		$post_meta_key = $this->get_postmeta_key( $term_id, $term->type );
+	public static function delete_editorial_metadata_term( int $term_id ): bool|WP_Error {
+		$term = self::get_editorial_metadata_term_by( 'id', $term_id );
+		if ( is_wp_error( $term ) ) {
+			return $term;
+		} else if ( ! $term ) {
+			return new WP_Error( 'invalid', __( "Editorial metadata term doesn't exist.", 'vip-workflow' ) );
+		}
+
+		// Delete the post meta for the term
+		$post_meta_key = self::get_postmeta_key( $term->meta[ self::METADATA_TYPE_KEY ], $term_id );
 		delete_post_meta_by_key( $post_meta_key );
 
-		$result = wp_delete_term( $term_id, self::METADATA_TAXONOMY );
+		delete_term_meta( $term_id, self::METADATA_TYPE_KEY );
 
+		delete_term_meta( $term_id, self::METADATA_POSTMETA_KEY );
+
+		$result = wp_delete_term( $term_id, self::METADATA_TAXONOMY );
 		if ( ! $result ) {
 			return new WP_Error( 'invalid', __( 'Unable to delete editorial metadata term.', 'vip-workflow' ) );
 		}
 
+		do_action( 'vw_editorial_metadata_term_deleted', $term_id );
+
 		// Reset the internal object cache
-		$this->editorial_metadata_terms_cache = array();
-
-		// Re-order the positions after deletion
-		$editorial_metadata_terms = $this->get_editorial_metadata_terms();
-
-		// ToDo: Optimize this to only work on the next or previous item.
-		$current_postition = 1;
-
-		// save each status with the new position
-		foreach ( $editorial_metadata_terms as $editorial_metadata_term ) {
-			$this->update_editorial_metadata_term( $editorial_metadata_term->term_id, [ 'position' => $current_postition ] );
-
-			$current_postition++;
-		}
+		self::$editorial_metadata_terms_cache = [];
 
 		return $result;
 	}
 
-	 /** Generate the args for registering post meta
+	/** Generate the args for registering post meta
 	 *
 	 * @param WP_Term $term The term object
 	 * @return array $args Post meta args
 	 */
-	public function get_postmeta_args( $term ) {
+	public static function get_postmeta_args( WP_Term $term ): array {
 		$arg_type = '';
-		switch ( $term->type ) {
+		switch ( $term->meta[ self::METADATA_TYPE_KEY ] ) {
 			case 'checkbox':
 				$arg_type = 'boolean';
 				break;
+			case 'date':
 			case 'text':
 				$arg_type = 'string';
 				break;
 		}
+
 		$args = [
-			'type' => $arg_type,
-			'description' => $term->description,
-			'single' => true,
-			'show_in_rest' => true,
+			'type'              => $arg_type,
+			'description'       => $term->description,
+			'single'            => true,
+			'show_in_rest'      => true,
 			'sanitize_callback' => function ( $value ) use ( $arg_type ) {
 				switch ( $arg_type ) {
 					case 'boolean':
@@ -441,6 +447,7 @@ class Editorial_Metadata extends Module {
 				}
 			},
 		];
+
 		return $args;
 	}
 
@@ -449,17 +456,16 @@ class Editorial_Metadata extends Module {
 	 *
 	 * Key is in the form of vw_editorial_meta_{type}_{term_id}
 	 *
-	 * @param WP_Term $term The term object
+	 * @param string $term_type The type of term
+	 * @param int $term_id The term's unique ID
 	 * @return string $postmeta_key Unique key
 	 */
-	public function get_postmeta_key( $term ) {
-		$key = self::METADATA_POSTMETA_KEY;
-		$prefix = "{$key}_{$term->type}";
-		$postmeta_key = "{$prefix}_" . $term->term_id;
+	public static function get_postmeta_key( string $term_type, int $term_id ): string {
+		$key          = self::METADATA_TAXONOMY;
+		$prefix       = "{$key}_{$term_type}";
+		$postmeta_key = "{$prefix}_" . $term_id;
 		return $postmeta_key;
 	}
-
-	public function print_configure_view() {
-		include_once __DIR__ . '/views/manage-editorial-metadata.php';
-	}
 }
+
+EditorialMetadata::init();
