@@ -1,249 +1,216 @@
 <?php
 /**
- * class Custom_Status
+ * class CustomStatus
  * Custom statuses make it simple to define the different stages in your publishing workflow.
  */
 
 namespace VIPWorkflow\Modules;
 
-require_once __DIR__ . '/rest/custom-status.php';
+// REST endpoints
+require_once __DIR__ . '/rest/custom-status-endpoint.php';
 
-use VIPWorkflow\VIP_Workflow;
-use VIPWorkflow\Modules\Shared\PHP\Module;
-use VIPWorkflow\Modules\CustomStatus\REST\EditStatus;
+// Term meta
+require_once __DIR__ . '/meta/required-user-id-handler.php';
+require_once __DIR__ . '/meta/required-metadata-id-handler.php';
+require_once __DIR__ . '/meta/position-handler.php';
+
+use VIPWorkflow\Modules\Shared\PHP\InstallUtilities;
+use VIPWorkflow\Modules\Shared\PHP\OptionsUtilities;
+use VIPWorkflow\Modules\CustomStatus\REST\CustomStatusEndpoint;
+use VIPWorkflow\Modules\Shared\PHP\HelperUtilities;
+
 use WP_Error;
 use WP_Query;
-use function VIPWorkflow\Modules\Shared\PHP\_vw_wp_link_page;
+use WP_Term;
+use WP_Post;
 
-class Custom_Status extends Module {
-
-	public $module;
-
-	private $custom_statuses_cache = [];
+class CustomStatus {
 
 	// This is taxonomy name used to store all our custom statuses
-	const TAXONOMY_KEY = 'post_status';
+	const TAXONOMY_KEY = 'vw_post_status';
 
-	/**
-	 * Register the module with VIP Workflow but don't do anything else
-	 */
-	public function __construct() {
+	const SETTINGS_SLUG = 'vw-custom-status';
 
-		$this->module_url = $this->get_module_url( __FILE__ );
-		// Register the module with VIP Workflow
-		$args         = [
-			'title'                => __( 'Workflow Config', 'vip-workflow' ),
-			'short_description'    => __( 'Configure your editorial workflow.', 'vip-workflow' ),
-			'extended_description' => __( 'Starting from the top, each post status represents the publishing worklow to be followed. This workflow can be configured by re-ordering statuses as well as editing/deleting and creating new ones.', 'vip-workflow' ),
-			'module_url'           => $this->module_url,
-			'slug'                 => 'custom-status',
-			'configure_page_cb'    => 'print_configure_view',
-		];
-		$this->module = VIP_Workflow::instance()->register_module( 'custom_status', $args );
-	}
+	// The metadata keys for the custom status term
+	const METADATA_POSITION_KEY          = 'position';
+	const METADATA_REQ_EDITORIAL_IDS_KEY = 'required_metadata_ids';
+	const METADATA_REQ_EDITORIALS_KEY    = 'required_metadatas';
+	const METADATA_REQ_USER_IDS_KEY      = 'required_user_ids';
+	const METADATA_REQ_USERS_KEY         = 'required_users';
 
-	/**
-	 * Initialize the Custom_Status class if the module is active
-	 */
-	public function init() {
-		// Register custom statuses as a taxonomy
-		$this->register_custom_statuses();
+	private static $custom_statuses_cache = [];
+
+	public static function init(): void {
+		// Register the taxonomy we use with WordPress core, and ensure it's registered after editorial metadata
+		add_action( 'init', [ __CLASS__, 'register_custom_status_taxonomy' ] );
+
+		// Register the custom statuses in core
+		add_action( 'init', [ __CLASS__, 'register_custom_statuses' ] );
+
+		// Setup custom statuses on first install
+		add_action( 'init', [ __CLASS__, 'setup_install' ] );
 
 		// Register our settings
-		if ( ! $this->disable_custom_statuses_for_post_type() ) {
-			// Load CSS and JS resources that we probably need in the admin page
-			add_action( 'admin_enqueue_scripts', [ $this, 'action_admin_enqueue_scripts' ] );
+		if ( ! HelperUtilities::is_current_post_type_unsupported() ) {
+			// Load CSS and JS resources for the admin page
+			add_action( 'admin_enqueue_scripts', [ __CLASS__, 'action_admin_enqueue_scripts' ] );
 
 			// Assets for block editor UI.
-			add_action( 'enqueue_block_editor_assets', [ $this, 'load_scripts_for_block_editor' ] );
+			add_action( 'enqueue_block_editor_assets', [ __CLASS__, 'load_scripts_for_block_editor' ] );
 
 			// Assets for iframed block editor and editor UI.
-			add_action( 'enqueue_block_editor_assets', [ $this, 'load_styles_for_block_editor' ] );
+			add_action( 'enqueue_block_editor_assets', [ __CLASS__, 'load_styles_for_block_editor' ] );
 		}
 
-		add_action( 'admin_print_scripts', [ $this, 'post_admin_header' ] );
+		add_action( 'admin_print_scripts', [ __CLASS__, 'post_admin_header' ] );
 
 		// Add custom statuses to the post states.
-		add_filter( 'display_post_states', [ $this, 'add_status_to_post_states' ], 10, 2 );
+		add_filter( 'display_post_states', [ __CLASS__, 'add_status_to_post_states' ], 10, 2 );
 
-		// These seven-ish methods are hacks for fixing bugs in WordPress core
-		add_action( 'admin_init', [ $this, 'check_timestamp_on_publish' ] );
-		add_filter( 'wp_insert_post_data', [ $this, 'fix_custom_status_timestamp' ], 10, 2 );
-		add_filter( 'wp_insert_post_data', [ $this, 'maybe_keep_post_name_empty' ], 10, 2 );
-		add_filter( 'pre_wp_unique_post_slug', [ $this, 'fix_unique_post_slug' ], 10, 6 );
-		add_filter( 'preview_post_link', [ $this, 'fix_preview_link_part_one' ] );
-		add_filter( 'post_link', [ $this, 'fix_preview_link_part_two' ], 10, 3 );
-		add_filter( 'page_link', [ $this, 'fix_preview_link_part_two' ], 10, 3 );
-		add_filter( 'post_type_link', [ $this, 'fix_preview_link_part_two' ], 10, 3 );
-		add_filter( 'preview_post_link', [ $this, 'fix_preview_link_part_three' ], 11, 2 );
-		add_filter( 'get_sample_permalink', [ $this, 'fix_get_sample_permalink' ], 10, 5 );
-		add_filter( 'get_sample_permalink_html', [ $this, 'fix_get_sample_permalink_html' ], 10, 5 );
-		add_filter( 'post_row_actions', [ $this, 'fix_post_row_actions' ], 10, 2 );
-		add_filter( 'page_row_actions', [ $this, 'fix_post_row_actions' ], 10, 2 );
+		// Register sidebar menu
+		add_action( 'admin_menu', [ __CLASS__, 'add_admin_menu' ], 6 /* Prior to default registration of sub-pages */ );
 
-		// Pagination for custom post statuses when previewing posts
-		add_filter( 'wp_link_pages_link', [ $this, 'modify_preview_link_pagination_url' ], 10, 2 );
-
-		// REST endpoints
-		EditStatus::init();
-
-		add_filter( 'user_has_cap', [ $this, 'remove_or_add_publish_capability_for_user' ], 10, 3 );
+		// Add server-side controls to block post status movements that are prohihibited by workflow rules
+		add_filter( 'wp_insert_post_data', [ __CLASS__, 'maybe_block_post_update' ], 1000, 4 );
+		add_filter( 'user_has_cap', [ __CLASS__, 'remove_or_add_publish_capability_for_user' ], 10, 3 );
 	}
 
 	/**
-	 * Create the default set of custom statuses the first time the module is loaded
+	 * Register the post metadata taxonomy
+	 *
+	 * @access private
 	 */
-	public function install() {
+	public static function register_custom_status_taxonomy(): void {
+		// We need to make sure taxonomy is registered for all of the post types that support it
+		$supported_post_types = HelperUtilities::get_supported_post_types();
 
-		$default_terms = [
+		register_taxonomy( self::TAXONOMY_KEY, $supported_post_types,
 			[
-				'term' => __( 'Pitch', 'vip-workflow' ),
-				'args' => [
-					'slug'        => 'pitch',
-					'description' => __( 'Idea proposed; waiting for acceptance.', 'vip-workflow' ),
-					'position'    => 1,
-				],
-			],
-			[
-				'term' => __( 'Assigned', 'vip-workflow' ),
-				'args' => [
-					'slug'        => 'assigned',
-					'description' => __( 'Post idea assigned to writer.', 'vip-workflow' ),
-					'position'    => 2,
-				],
-			],
-			[
-				'term' => __( 'In Progress', 'vip-workflow' ),
-				'args' => [
-					'slug'        => 'in-progress',
-					'description' => __( 'Writer is working on the post.', 'vip-workflow' ),
-					'position'    => 3,
-				],
-			],
-			[
-				'term' => __( 'Draft', 'vip-workflow' ),
-				'args' => [
-					'slug'        => 'draft',
-					'description' => __( 'Post is a draft; not ready for review or publication.', 'vip-workflow' ),
-					'position'    => 4,
-				],
-			],
-			[
-				'term' => __( 'Pending Review' ),
-				'args' => [
-					'slug'        => 'pending',
-					'description' => __( 'Post needs to be reviewed by an editor.', 'vip-workflow' ),
-					'position'    => 5,
-				],
-			],
-		];
-
-		// Okay, now add the default statuses to the db if they don't already exist
-		foreach ( $default_terms as $term ) {
-			if ( ! term_exists( $term['term'], self::TAXONOMY_KEY ) ) {
-				$this->add_custom_status( $term['term'], $term['args'] );
-			}
-		}
-	}
-
-	/**
-	 * Makes the call to register_post_status to register the user's custom statuses.
-	 * Also unregisters draft and pending, in case the user doesn't want them.
-	 */
-	public function register_custom_statuses() {
-		global $wp_post_statuses;
-
-		if ( $this->disable_custom_statuses_for_post_type() ) {
-			return;
-		}
-
-		// Register new taxonomy so that we can store all our fancy new custom statuses (or is it stati?)
-		if ( ! taxonomy_exists( self::TAXONOMY_KEY ) ) {
-			$args = [
 				'hierarchical'          => false,
 				'update_count_callback' => '_update_post_term_count',
 				'label'                 => false,
 				'query_var'             => false,
 				'rewrite'               => false,
 				'show_ui'               => false,
-			];
-			register_taxonomy( self::TAXONOMY_KEY, 'post', $args );
-		}
-
-		if ( function_exists( 'register_post_status' ) ) {
-			// Users can delete draft and pending statuses if they want, so let's get rid of them
-			// They'll get re-added if the user hasn't "deleted" them
-			unset( $wp_post_statuses['draft'] );
-			unset( $wp_post_statuses['pending'] );
-
-			$custom_statuses = $this->get_custom_statuses();
-
-			// Unfortunately, register_post_status() doesn't accept a
-			// post type argument, so we have to register the post
-			// statuses for all post types. This results in
-			// all post statuses for a post type appearing at the top
-			// of manage posts if there is a post with the status
-			foreach ( $custom_statuses as $status ) {
-				register_post_status( $status->slug, [
-					'label'                     => $status->name,
-					'protected'                 => true,
-					'_builtin'                  => false,
-					'label_count'               => _n_noop( "{$status->name} <span class='count'>(%s)</span>", "{$status->name} <span class='count'>(%s)</span>" ),
-					'show_in_admin_status_list' => true,
-					'show_in_admin_all_list'    => true,
-				] );
-			}
-		}
+			]
+		);
 	}
 
 	/**
-	 * Whether custom post statuses should be disabled for this post type.
-	 * Used to stop custom statuses from being registered for post types that don't support them.
-	 *
-	 * @return bool
+	 * Makes the call to register_post_status to register the user's custom statuses.
+	 * Also unregisters pending, in case the user doesn't want them.
 	 */
-	public function disable_custom_statuses_for_post_type( $post_type = null ) {
-		global $pagenow;
+	public static function register_custom_statuses(): void {
+		$custom_statuses = self::get_custom_statuses();
 
-		// Only allow deregistering on 'edit.php' and 'post.php'
-		if ( ! in_array( $pagenow, [ 'edit.php', 'post.php', 'post-new.php' ] ) ) {
-			return false;
+		// Unfortunately, register_post_status() doesn't accept a
+		// post type argument, so we have to register the post
+		// statuses for all post types. This results in
+		// all post statuses for a post type appearing at the top
+		// of manage posts if there is a post with the status
+		foreach ( $custom_statuses as $status ) {
+			register_post_status( $status->slug, [
+				'label'                     => $status->name,
+				'protected'                 => true,
+				'_builtin'                  => false,
+				'label_count'               => _n_noop( "{$status->name} <span class='count'>(%s)</span>", "{$status->name} <span class='count'>(%s)</span>" ),
+				'show_in_admin_status_list' => true,
+				'show_in_admin_all_list'    => true,
+				'date_floating'             => true,
+			] );
 		}
-
-		if ( is_null( $post_type ) ) {
-			$post_type = $this->get_current_post_type();
-		}
-
-		if ( $post_type && ! in_array( $post_type, $this->get_supported_post_types() ) ) {
-			return true;
-		}
-
-		return false;
 	}
 
-	public function configure_page_cb() {
-		// do nothing
+
+	/**
+	 * Load default custom statuses the first time the module is loaded
+	 *
+	 * @access private
+	 */
+	public static function setup_install(): void {
+		InstallUtilities::install_if_first_run( self::SETTINGS_SLUG, function () {
+			$default_terms = [
+				[
+					'name'        => __( 'Pitch', 'vip-workflow' ),
+					'slug'        => 'pitch',
+					'description' => __( 'Idea proposed; waiting for acceptance.', 'vip-workflow' ),
+					'position'    => 1,
+				],
+				[
+					'name'        => __( 'Assigned', 'vip-workflow' ),
+					'slug'        => 'assigned',
+					'description' => __( 'Post idea assigned to writer.', 'vip-workflow' ),
+					'position'    => 2,
+				],
+				[
+					'name'        => __( 'In Progress', 'vip-workflow' ),
+					'slug'        => 'in-progress',
+					'description' => __( 'Writer is working on the post.', 'vip-workflow' ),
+					'position'    => 3,
+				],
+				[
+					'name'        => __( 'Draft', 'vip-workflow' ),
+					'slug'        => 'draft',
+					'description' => __( 'Post is a draft; not ready for review or publication.', 'vip-workflow' ),
+					'position'    => 4,
+				],
+				[
+					'name'        => __( 'Pending Review' ),
+					'slug'        => 'pending',
+					'description' => __( 'Post needs to be reviewed by an editor.', 'vip-workflow' ),
+					'position'    => 5,
+				],
+			];
+
+			// Add the custom statuses if the slugs don't conflict
+			foreach ( $default_terms as $term ) {
+				if ( ! term_exists( $term['slug'], self::TAXONOMY_KEY ) ) {
+					self::add_custom_status( $term );
+				}
+			}
+		});
+	}
+
+	/**
+	 * Register admin sidebar menu
+	 *
+	 * @access private
+	 */
+	public static function add_admin_menu(): void {
+		$menu_title = __( 'VIP Workflow', 'vip-workflow' );
+
+		add_menu_page( $menu_title, $menu_title, 'manage_options', self::SETTINGS_SLUG, [ __CLASS__, 'render_settings_view' ] );
+	}
+
+	/**
+	 * Primary configuration page for custom status class, which is also the main entry point for configuring the plugin
+	 */
+	public static function render_settings_view(): void {
+		include_once __DIR__ . '/views/manage-workflow.php';
 	}
 
 	/**
 	 * Enqueue resources that we need in the admin settings page
+	 *
+	 * @access private
 	 */
-	public function action_admin_enqueue_scripts() {
+	public static function action_admin_enqueue_scripts(): void {
 		// Load Javascript we need to use on the configuration views
-		if ( $this->is_whitelisted_settings_view() ) {
+		if ( HelperUtilities::is_settings_view_loaded( self::SETTINGS_SLUG ) ) {
 			$asset_file = include VIP_WORKFLOW_ROOT . '/dist/modules/custom-status/custom-status-configure.asset.php';
 			wp_enqueue_script( 'vip-workflow-custom-status-configure', VIP_WORKFLOW_URL . 'dist/modules/custom-status/custom-status-configure.js', $asset_file['dependencies'], $asset_file['version'], true );
 			wp_enqueue_style( 'vip-workflow-custom-status-styles', VIP_WORKFLOW_URL . 'dist/modules/custom-status/custom-status-configure.css', [ 'wp-components' ], $asset_file['version'] );
 
 			wp_localize_script( 'vip-workflow-custom-status-configure', 'VW_CUSTOM_STATUS_CONFIGURE', [
-				'custom_statuses'    => $this->get_custom_statuses(),
-				'url_edit_status'    => EditStatus::get_crud_url(),
-				'url_reorder_status' => EditStatus::get_reorder_url(),
+				'custom_statuses'     => self::modify_custom_statuses_with_editorial_metadata(),
+				'editorial_metadatas' => EditorialMetadata::get_editorial_metadata_terms(),
+				'url_edit_status'     => CustomStatusEndpoint::get_crud_url(),
+				'url_reorder_status'  => CustomStatusEndpoint::get_reorder_url(),
 			] );
 		}
 
 		// Custom javascript to modify the post status dropdown where it shows up
-		if ( $this->is_whitelisted_page() ) {
+		if ( self::is_whitelisted_page() ) {
 			$asset_file   = include VIP_WORKFLOW_ROOT . '/dist/modules/custom-status/custom-status.asset.php';
 			$dependencies = [ ...$asset_file['dependencies'], 'jquery', 'post' ];
 			wp_enqueue_script( 'vip_workflow-custom_status', VIP_WORKFLOW_URL . 'dist/modules/custom-status/custom-status.js', $dependencies, $asset_file['version'], true );
@@ -260,38 +227,78 @@ class Custom_Status extends Module {
 		}
 	}
 
-	public function load_scripts_for_block_editor() {
+	/**
+	 * Enqueue resources that we need in the admin settings page
+	 *
+	 * @access private
+	 */
+	public static function load_scripts_for_block_editor(): void {
 		$asset_file = include VIP_WORKFLOW_ROOT . '/dist/modules/custom-status/custom-status-block.asset.php';
 		wp_enqueue_script( 'vip-workflow-block-custom-status-script', VIP_WORKFLOW_URL . 'dist/modules/custom-status/custom-status-block.js', $asset_file['dependencies'], $asset_file['version'], true );
 
-		$publish_guard_enabled = ( 'on' === VIP_Workflow::instance()->settings->module->options->publish_guard ) ? true : false;
+		$publish_guard_enabled = ( 'on' === OptionsUtilities::get_options_by_key( 'publish_guard' ) ) ? true : false;
 
 		wp_localize_script( 'vip-workflow-block-custom-status-script', 'VW_CUSTOM_STATUSES', [
+			'current_user_id'          => get_current_user_id(),
 			'is_publish_guard_enabled' => $publish_guard_enabled,
-			'status_terms'             => $this->get_custom_statuses(),
-			'supported_post_types'     => $this->get_supported_post_types(),
+			'status_terms'             => self::modify_custom_statuses_with_editorial_metadata(),
+			'supported_post_types'     => HelperUtilities::get_supported_post_types(),
 		] );
 	}
 
-	public function load_styles_for_block_editor() {
+	/**
+	 * Modify the custom statuses to include the editorial metadatas for UI purposes.
+	 *
+	 * This isn't done anywhere else due to the taxonomies being registered at different times.
+	 * In addition, registering the taxonomies in the wrong order can cause the manage posts page to break
+	 * as well as the default status for a post itself.
+	 *
+	 * @return array $custom_statuses The custom statuses with the editorial metadatas included
+	 */
+	private static function modify_custom_statuses_with_editorial_metadata(): array {
+		// map the editorial metadatas to their respective term_id so the term_id can be used to get the full object quickly.
+		$editorial_metadatas = EditorialMetadata::get_editorial_metadata_terms();
+		$editorial_metadatas = array_combine( array_column( $editorial_metadatas, 'term_id' ), $editorial_metadatas );
+
+		$custom_statuses = self::get_custom_statuses();
+
+		// Add the required editorial metadata to the custom statuses for UI purposes
+		foreach ( $custom_statuses as $status ) {
+			$required_metadata_ids = $status->meta[ self::METADATA_REQ_EDITORIAL_IDS_KEY ] ?? [];
+			$required_metadatas    = [];
+			foreach ( $required_metadata_ids as $metadata_id ) {
+				$required_metadatas[] = $editorial_metadatas[ $metadata_id ];
+			}
+			$status->meta[ self::METADATA_REQ_EDITORIALS_KEY ] = $required_metadatas;
+		}
+
+		return $custom_statuses;
+	}
+
+	/**
+	 * Enqueue resources that we need in the block editor
+	 *
+	 * @access private
+	 */
+	public static function load_styles_for_block_editor(): void {
 		$asset_file = include VIP_WORKFLOW_ROOT . '/dist/modules/custom-status/custom-status-block.asset.php';
 
-		wp_enqueue_style( 'vip-workflow-manager-styles', VIP_WORKFLOW_URL . 'dist/modules/custom-status/custom-status-block.css', [ 'wp-components' ], $asset_file['version'] );
+		wp_enqueue_style( 'vip-workflow-custom-status-styles', VIP_WORKFLOW_URL . 'dist/modules/custom-status/custom-status-block.css', [], $asset_file['version'] );
 	}
 
 	/**
 	 * Check whether custom status stuff should be loaded on this page
-	 *
-	 * @todo migrate this to the base module class
 	 */
-	public function is_whitelisted_page() {
+	public static function is_whitelisted_page(): bool {
 		global $pagenow;
 
-		if ( ! in_array( $this->get_current_post_type(), $this->get_supported_post_types() ) ) {
+		$current_post_type = HelperUtilities::get_current_post_type();
+
+		if ( ! in_array( $current_post_type, HelperUtilities::get_supported_post_types() ) ) {
 			return false;
 		}
 
-		$post_type_obj = get_post_type_object( $this->get_current_post_type() );
+		$post_type_obj = get_post_type_object( $current_post_type );
 
 		if ( ! current_user_can( $post_type_obj->cap->edit_posts ) ) {
 			return false;
@@ -303,27 +310,24 @@ class Custom_Status extends Module {
 
 	/**
 	 * Adds all necessary javascripts to make custom statuses work
-	 *
-	 * @todo Support private and future posts on edit.php view
 	 */
-	public function post_admin_header() {
+	public static function post_admin_header(): void {
 		global $post, $pagenow;
 
-		if ( $this->disable_custom_statuses_for_post_type() ) {
+		if ( HelperUtilities::is_current_post_type_unsupported() ) {
 			return;
 		}
 
-		// Get current user
+		// Set the current user, so we can check if they can publish posts
 		wp_get_current_user();
 
 		// Only add the script to Edit Post and Edit Page pages -- don't want to bog down the rest of the admin with unnecessary javascript
-		if ( $this->is_whitelisted_page() ) {
+		if ( self::is_whitelisted_page() ) {
 
-			$custom_statuses = $this->get_custom_statuses();
+			$custom_statuses = self::get_custom_statuses();
 
 			// $selected can be empty, but must be set because it's used as a JS variable
-			$selected      = '';
-			$selected_name = '';
+			$selected = '';
 
 			if ( ! empty( $post ) ) {
 				// Get the status of the current post
@@ -331,13 +335,6 @@ class Custom_Status extends Module {
 					$selected = $custom_statuses[0]->slug;
 				} else {
 					$selected = $post->post_status;
-				}
-
-				// Get the label of current status
-				foreach ( $custom_statuses as $status ) {
-					if ( $status->slug === $selected ) {
-						$selected_name = $status->name;
-					}
 				}
 			}
 
@@ -370,18 +367,59 @@ class Custom_Status extends Module {
 				];
 			}
 
-			$post_type_obj = get_post_type_object( $this->get_current_post_type() );
+			$post_type_obj = get_post_type_object( HelperUtilities::get_current_post_type() );
 
 			// Now, let's print the JS vars
 			?>
 				<script type="text/javascript">
 					var custom_statuses = <?php echo json_encode( $all_statuses ); ?>;
 					var current_status = '<?php echo esc_js( $selected ); ?>';
-					var current_status_name = '<?php echo esc_js( $selected_name ); ?>';
 					var current_user_can_publish_posts = <?php echo current_user_can( $post_type_obj->cap->publish_posts ) ? 1 : 0; ?>;
 				</script>
 			<?php
 		}
+	}
+
+	/**
+	 * Remove the ability to transition a post if custom status requires review from a nonpresent user
+	 *
+	 * @param array $data Post data submitted for update
+	 *
+	 * @return array $allcaps All capabilities for the user
+	 */
+	public static function maybe_block_post_update( array $data ): array|bool {
+		$status_slugs = wp_list_pluck( self::get_custom_statuses(), 'slug' );
+
+		// Ignore if it's not a post status and post type we support
+		if ( ! in_array( $data['post_type'], HelperUtilities::get_supported_post_types() ) ) {
+			return $data;
+		}
+
+		$new_post_status   = $data['post_status'];
+		$post_status_index = array_search( $new_post_status, $status_slugs, /* strict */ true );
+
+		if ( false === $post_status_index ) {
+			// This is not a supported custom status
+			return $data;
+		}
+
+		$prior_post_status_slug = $post_status_index > 0 ? $status_slugs[ $post_status_index - 1 ] : false;
+
+		if ( false === $prior_post_status_slug ) {
+			// This is the first custom status, so we don't need to block it
+			return $data;
+		}
+
+		$prior_post_status = self::get_custom_status_by( 'slug', $prior_post_status_slug );
+		$required_user_ids = $prior_post_status->meta[ self::METADATA_REQ_USER_IDS_KEY ] ?? [];
+
+		if ( $required_user_ids && ! in_array( get_current_user_id(), $required_user_ids, true ) ) {
+			// This status requires review, and the current user is not permitted to transition the post.
+			// Return an empty array, which will cause the update to fail with an error.
+			return false;
+		}
+
+		return $data;
 	}
 
 	/**
@@ -393,20 +431,21 @@ class Custom_Status extends Module {
 	 *
 	 * @return array $allcaps All capabilities for the user
 	 */
-	public function remove_or_add_publish_capability_for_user( $allcaps, $cap, $args ) {
+	public static function remove_or_add_publish_capability_for_user( array $allcaps, array $cap, array $args ): array {
 		global $post;
+
 		$supported_publish_caps_map = [
 			'post' => 'publish_posts',
 			'page' => 'publish_pages',
 		];
 
 		// Bail early if publish guard is off, or the post is already published, or the post is not available
-		if ( 'off' === VIP_Workflow::instance()->settings->module->options->publish_guard || ! $post || 'publish' === $post->post_status ) {
+		if ( ! $post || 'off' === OptionsUtilities::get_options_by_key( 'publish_guard' ) || 'publish' === $post->post_status ) {
 			return $allcaps;
 		}
 
 		// Bail early if the post type is not supported or if its a not supported capability for this guard
-		if ( ! in_array( $post->post_type, $this->get_supported_post_types() ) || ! isset( $supported_publish_caps_map[ $post->post_type ] ) ) {
+		if ( ! in_array( $post->post_type, HelperUtilities::get_supported_post_types() ) || ! isset( $supported_publish_caps_map[ $post->post_type ] ) ) {
 			return $allcaps;
 		}
 
@@ -416,7 +455,7 @@ class Custom_Status extends Module {
 			return $allcaps;
 		}
 
-		$custom_statuses = VIP_Workflow::instance()->custom_status->get_custom_statuses();
+		$custom_statuses = self::get_custom_statuses();
 		$status_slugs    = wp_list_pluck( $custom_statuses, 'slug' );
 
 		// Bail early if the post is not using a custom status
@@ -426,16 +465,29 @@ class Custom_Status extends Module {
 
 		$status_before_publish = $custom_statuses[ array_key_last( $custom_statuses ) ];
 
-		// If the post status is not the last status, remove the publish capability or else add it back in
+		// Ensure publishing is disabled for all but the last status
 		if ( $status_before_publish->slug !== $post->post_status ) {
-			// Remove the publish capability based on the post type
 			$allcaps[ $supported_publish_caps_map[ $post->post_type ] ] = false;
-		} else {
-			// Remove the publish capability based on the post type
-			$allcaps[ $supported_publish_caps_map[ $post->post_type ] ] = true;
 		}
 
 		return $allcaps;
+	}
+
+	/**
+	 * Perform the necessary data cleanup when an error occurs while saving the custom status,
+	 * and generate a WP_Error object.
+	 *
+	 * @param integer $term_id The ID of the term that failed to save
+	 * @param string $meta The metadata that failed to save
+	 * @return WP_Error The WP_Error object
+	 */
+	private static function generate_error_and_delete_bad_data( int $term_id, string $meta ): WP_Error {
+		// Trigger the deletion of the metadata associated with the status
+		do_action( 'vw_delete_custom_status_meta', $term_id );
+		wp_delete_term( $term_id, self::TAXONOMY_KEY );
+
+		/* translators: %s: meta key that failed to save */
+		return new WP_Error( 'invalid', sprintf( __( 'Unable to create the custom status, as the %s failed to save.', 'vip-workflow' ), $meta ) );
 	}
 
 	/**
@@ -445,54 +497,81 @@ class Custom_Status extends Module {
 	 * The arguments decide how the term is handled based on the $args parameter.
 	 * The following is a list of the available overrides and the defaults.
 	 *
+	 * 'slug'. Expected to be a string. There is no default.
+	 *
 	 * 'description'. There is no default. If exists, will be added to the database
 	 * along with the term. Expected to be a string.
 	 *
-	 * 'slug'. Expected to be a string. There is no default.
+	 * 'required_user_ids'. An optional array of user IDs that are required to review the post in the current status.
 	 *
 	 * @param int|string $term The status to add or update
 	 * @param array|string $args Change the values of the inserted term
 	 *
 	 * @return object|WP_Error $inserted_term The newly inserted term object or a WP_Error object
 	 */
-	public function add_custom_status( $term, $args = [] ) {
-		// Term is always added to the end of the list
-		$default_position = count( $this->get_custom_statuses() ) + 1;
-
-		$slug = ( ! empty( $args['slug'] ) ) ? $args['slug'] : sanitize_title( $term );
-		unset( $args['slug'] );
-
+	public static function add_custom_status( array $args ): WP_Term|WP_Error {
 		if ( ! isset( $args['position'] ) ) {
-			$args['position'] = $default_position;
+			// get the existing statuses, ordered by position
+			$custom_statuses = self::get_custom_statuses();
+
+			// get the last status position
+			$last_position = $custom_statuses[ array_key_last( $custom_statuses ) ]->meta[ self::METADATA_POSITION_KEY ];
+
+			// set the new status position to be one more than the last status
+			$args['position'] = $last_position + 1;
+		}
+
+		$term_to_save = [
+			'slug'        => $args['slug'] ?? sanitize_title( $args['name'] ),
+			'description' => $args['description'] ?? '',
+		];
+
+		$term_name = $args['name'];
+
+		$inserted_term = wp_insert_term( $term_name, self::TAXONOMY_KEY, $term_to_save );
+
+		if ( is_wp_error( $inserted_term ) ) {
+			return $inserted_term;
 		}
 
 		/**
-		 * Fires before a custom status is added to the database.
+		 * Fires after a custom status is added to the database.
 		 *
 		 * @param string $term The status to add or update
 		 * @param string $slug The slug of the status
 		 * @param string $description Change the values of the inserted term
 		 */
-		do_action( 'vw_add_custom_status', $term, $slug, $args['description'] );
-
-		$encoded_description = $this->get_encoded_description( $args );
-
-		$inserted_term = wp_insert_term( $term, self::TAXONOMY_KEY, [
-			'slug'        => $slug,
-			'description' => $encoded_description,
-		] );
+		do_action( 'vw_add_custom_status', $term_name, $term_to_save['slug'], $term_to_save['description'] );
 
 		// Reset our internal object cache
-		$this->custom_statuses_cache = [];
+		self::$custom_statuses_cache = [];
 
-		// Populate the inserted term with the new values, or else only the term_taxonomy_id and term_id are returned.
-		if ( is_wp_error( $inserted_term ) ) {
-			return $inserted_term;
-		} else {
-			$inserted_term = $this->get_custom_status_by( 'id', $inserted_term['term_id'] );
+		$term_id = $inserted_term['term_id'];
+
+		$position              = $args[ self::METADATA_POSITION_KEY ];
+		$required_metadata_ids = $args[ self::METADATA_REQ_EDITORIAL_IDS_KEY ] ?? [];
+		$required_user_ids     = $args[ self::METADATA_REQ_USER_IDS_KEY ] ?? [];
+
+		// In case of failure, data cleanup happens which includes the term and the meta keys.
+
+		$position_meta_result = update_term_meta( $term_id, self::METADATA_POSITION_KEY, $position );
+		if ( is_wp_error( $position_meta_result ) ) {
+			return self::generate_error_and_delete_bad_data( $term_id, 'position' );
 		}
 
-		return $inserted_term;
+		$required_metadata_ids_result = update_term_meta( $term_id, self::METADATA_REQ_EDITORIAL_IDS_KEY, $required_metadata_ids );
+		if ( is_wp_error( $required_metadata_ids_result ) ) {
+			return self::generate_error_and_delete_bad_data( $term_id, 'required editorial metadata fields' );
+		}
+
+		$required_user_ids_result = update_term_meta( $term_id, self::METADATA_REQ_USER_IDS_KEY, $required_user_ids );
+		if ( is_wp_error( $required_user_ids_result ) ) {
+			return self::generate_error_and_delete_bad_data( $term_id, 'required users' );
+		}
+
+		$term_result = self::get_custom_status_by( 'id', $term_id );
+
+		return $term_result;
 	}
 
 	/**
@@ -502,23 +581,26 @@ class Custom_Status extends Module {
 	 * @param array $args Any arguments to be updated
 	 * @return object $updated_status Newly updated status object
 	 */
-	public function update_custom_status( $status_id, $args = [] ) {
-		$old_status = $this->get_custom_status_by( 'id', $status_id );
-		if ( ! $old_status ) {
+	public static function update_custom_status( int $status_id, array $args = [] ): WP_Term|WP_Error {
+		$old_status = self::get_custom_status_by( 'id', $status_id );
+		if ( is_wp_error( $old_status ) ) {
+			return $old_status;
+		} elseif ( ! $old_status ) {
 			return new WP_Error( 'invalid', __( "Custom status doesn't exist.", 'vip-workflow' ) );
 		}
 
 		// Reset our internal object cache
-		$this->custom_statuses_cache = [];
+		self::$custom_statuses_cache = [];
 
 		// Prevent user from changing draft name or slug
-		if ( 'draft' === $old_status->slug
+		if ( self::is_restricted_status( $old_status->slug )
 		&& (
 			( isset( $args['name'] ) && $args['name'] !== $old_status->name )
 			||
 			( isset( $args['slug'] ) && $args['slug'] !== $old_status->slug )
 		) ) {
-			return new WP_Error( 'invalid', __( 'Changing the name and slug of "Draft" is not allowed', 'vip-workflow' ) );
+			// translators: %s: Post status, like "Draft"
+			return new WP_Error( 'restricted', sprintf( __( 'Changing the name and slug of a restricted status (%s) is not allowed.', 'vip-workflow' ), $old_status->name ) );
 		}
 
 		// If the name was changed, we need to change the slug
@@ -527,23 +609,55 @@ class Custom_Status extends Module {
 		}
 
 		// Reassign posts to new status slug if the slug changed and isn't restricted
-		if ( isset( $args['slug'] ) && $args['slug'] != $old_status->slug && ! $this->is_restricted_status( $old_status->slug ) ) {
+		if ( isset( $args['slug'] ) && $args['slug'] != $old_status->slug && ! self::is_restricted_status( $old_status->slug ) ) {
 			$new_status        = $args['slug'];
-			$reassigned_result = $this->reassign_post_status( $old_status->slug, $new_status );
+			$reassigned_result = self::reassign_post_status( $old_status->slug, $new_status );
 			// If the reassignment failed, return the error
 			if ( is_wp_error( $reassigned_result ) ) {
 				return $reassigned_result;
 			}
 		}
-		// We're encoding metadata that isn't supported by default in the term's description field
-		$args_to_encode                = [];
-		$args_to_encode['description'] = ( isset( $args['description'] ) ) ? $args['description'] : $old_status->description;
-		$args_to_encode['position']    = ( isset( $args['position'] ) ) ? $args['position'] : $old_status->position;
-		$encoded_description           = $this->get_encoded_description( $args_to_encode );
-		$args['description']           = $encoded_description;
+
+		$term_fields_to_update = [
+			'name'        => isset( $args['name'] ) ? $args['name'] : $old_status->name,
+			'slug'        => isset( $args['slug'] ) ? $args['slug'] : $old_status->slug,
+			'description' => isset( $args['description'] ) ? $args['description'] : $old_status->description,
+		];
+
+		// Update the metadata first, as if it fails we don't want to update the term
+
+		if ( isset( $args[ self::METADATA_POSITION_KEY ] ) ) {
+			$position_meta_result = update_term_meta( $status_id, self::METADATA_POSITION_KEY, $args[ self::METADATA_POSITION_KEY ] );
+			if ( is_wp_error( $position_meta_result ) ) {
+				return new WP_Error( 'invalid', __( 'Unable to update custom status, as the position failed to save.', 'vip-workflow' ) );
+			}
+		}
+
+		if ( isset( $args[ self::METADATA_REQ_EDITORIAL_IDS_KEY ] ) ) {
+			$required_metadata_ids_result = update_term_meta( $status_id, self::METADATA_REQ_EDITORIAL_IDS_KEY, $args[ self::METADATA_REQ_EDITORIAL_IDS_KEY ] );
+			if ( is_wp_error( $required_metadata_ids_result ) ) {
+				return new WP_Error( 'invalid', __( 'Unable to update custom status, as the required editorial metadata fields failed to save.', 'vip-workflow' ) );
+			}
+		}
+
+		if ( isset( $args[ self::METADATA_REQ_USER_IDS_KEY ] ) ) {
+			$req_user_ids_meta_result = update_term_meta( $status_id, self::METADATA_REQ_USER_IDS_KEY, $args[ self::METADATA_REQ_USER_IDS_KEY ] );
+			if ( is_wp_error( $req_user_ids_meta_result ) ) {
+				return new WP_Error( 'invalid', __( 'Unable to update custom status, as the required users failed to save.', 'vip-workflow' ) );
+			}
+		}
+
+		$updated_term = wp_update_term( $status_id, self::TAXONOMY_KEY, $term_fields_to_update );
+
+		// Reset status cache again, as reassign_post_status() will recache prior statuses
+		self::$custom_statuses_cache = [];
+
+		if ( is_wp_error( $updated_term ) ) {
+			return $updated_term;
+		}
 
 		/**
-		 * Fires before a custom status is updated in the database.
+		 * Fires after a custom status is updated in the database.
 		 *
 		 * @param int $status_id The ID of the status being updated
 		 * @param string $slug The slug of the status being updated
@@ -556,19 +670,9 @@ class Custom_Status extends Module {
 			isset( $args['position'] ) ? $args['position'] : $old_status->position
 		);
 
-		$updated_status = wp_update_term( $status_id, self::TAXONOMY_KEY, $args );
+		$status_result = self::get_custom_status_by( 'id', $status_id );
 
-		// Reset status cache again, as reassign_post_status() will recache prior statuses
-		$this->custom_statuses_cache = [];
-
-		// Populate the updated term with the new values, or else only the term_taxonomy_id and term_id are returned.
-		if ( is_wp_error( $updated_status ) ) {
-			return $updated_status;
-		} else {
-			$updated_status = $this->get_custom_status_by( 'id', $status_id );
-		}
-
-		return $updated_status;
+		return $status_result;
 	}
 
 	/**
@@ -577,64 +681,69 @@ class Custom_Status extends Module {
 	 * Partly a wrapper for the wp_delete_term function.
 	 * BUT, also reassigns posts that currently have the deleted status assigned.
 	 */
-	public function delete_custom_status( $status_id, $args = [] ) {
+	public static function delete_custom_status( int $status_id ): bool|WP_Error {
 		// Get slug for the old status
-		$old_status = $this->get_custom_status_by( 'id', $status_id );
-		if ( ! $old_status ) {
+		$old_status = self::get_custom_status_by( 'id', $status_id );
+		if ( is_wp_error( $old_status ) ) {
+			return $old_status;
+		} elseif ( ! $old_status ) {
 			return new WP_Error( 'invalid', __( "Custom status doesn't exist.", 'vip-workflow' ) );
 		}
 
 		$old_status_slug = $old_status->slug;
 
-		if ( $this->is_restricted_status( $old_status_slug ) || 'draft' === $old_status_slug ) {
-			return new WP_Error( 'restricted', __( 'Restricted status ', 'vip-workflow' ) . '(' . $old_status->name . ')' );
+		if ( self::is_restricted_status( $old_status_slug ) ) {
+			// translators: %s: Post status, like "Draft"
+			return new WP_Error( 'restricted', sprintf( __( 'Restricted status (%s) cannot be deleted.', 'vip-workflow' ), $old_status->name ) );
 		}
 
 		// Reset our internal object cache
-		$this->custom_statuses_cache = [];
+		self::$custom_statuses_cache = [];
 
 		// Get the new status to reassign posts to, which would be the first custom status.
 		// In the event that the first custom status is being deleted, we'll reassign to the second custom status.
-		// Since draft cannot be deleted, we don't need to worry about ever getting index out of bounds.
-		$custom_statuses = $this->get_custom_statuses();
+		// Since draft and pending review cannot be deleted, we don't need to worry about ever getting index out of bounds.
+		$custom_statuses = self::get_custom_statuses();
 		$new_status_slug = $custom_statuses[0]->slug;
 		if ( $old_status_slug === $new_status_slug ) {
 			$new_status_slug = $custom_statuses[1]->slug;
 		}
 
-		$reassigned_result = $this->reassign_post_status( $old_status_slug, $new_status_slug );
+		$reassigned_result = self::reassign_post_status( $old_status_slug, $new_status_slug );
 		// If the reassignment failed, return the error
 		if ( is_wp_error( $reassigned_result ) ) {
 			return $reassigned_result;
 		}
 
+		// Trigger the deletion of the metadata associated with the status
+		do_action( 'vw_delete_custom_status_meta', $status_id );
+
+		$result = wp_delete_term( $status_id, self::TAXONOMY_KEY );
+		if ( ! $result ) {
+			return new WP_Error( 'invalid', __( 'Unable to delete custom status.', 'vip-workflow' ) );
+		}
+
 		/**
-		 * Fires before a custom status is deleted from the database.
+		 * Fires after a custom status is deleted from the database.
 		 *
 		 * @param int $status_id The ID of the status being deleted
 		 * @param string $old_status_slug The slug of the status being deleted
 		 */
 		do_action( 'vw_delete_custom_status', $status_id, $old_status_slug );
 
-		$result = wp_delete_term( $status_id, self::TAXONOMY_KEY, $args );
-		if ( ! $result ) {
-			return new WP_Error( 'invalid', __( 'Unable to delete custom status.', 'vip-workflow' ) );
-		}
-
 		// Reset status cache again, as reassign_post_status() will recache prior statuses
-		$this->custom_statuses_cache = [];
+		self::$custom_statuses_cache = [];
 
 		// Re-order the positions after deletion
-		$custom_statuses = $this->get_custom_statuses();
+		$custom_statuses = self::get_custom_statuses();
 
-		// ToDo: Optimize this to only work on the next or previous item.
 		$current_postition = 1;
 
 		// save each status with the new position
 		foreach ( $custom_statuses as $status ) {
-			$this->update_custom_status( $status->term_id, [ 'position' => $current_postition ] );
+			self::update_custom_status( $status->term_id, [ 'position' => $current_postition ] );
 
-			$current_postition++;
+			++$current_postition;
 		}
 
 		return $result;
@@ -646,87 +755,95 @@ class Custom_Status extends Module {
 	 * @param array|string $statuses
 	 * @return array $statuses All of the statuses
 	 */
-	public function get_custom_statuses() {
-		if ( $this->disable_custom_statuses_for_post_type() ) {
-			return $this->get_core_post_statuses();
+	public static function get_custom_statuses(): array {
+		if ( HelperUtilities::is_current_post_type_unsupported() ) {
+			return self::get_core_statuses();
 		}
 
 		// Internal object cache for repeat requests
-		if ( ! empty( $this->custom_statuses_cache ) ) {
-			return $this->custom_statuses_cache;
+		if ( ! empty( self::$custom_statuses_cache ) ) {
+			return self::$custom_statuses_cache;
 		}
 
-		// Handle if the requested taxonomy doesn't exist
 		$statuses = get_terms( [
 			'taxonomy'   => self::TAXONOMY_KEY,
 			'hide_empty' => false,
+			'orderby'    => 'meta_value_num',
+			'order'      => 'ASC',
+			'meta_key'   => self::METADATA_POSITION_KEY,
 		]);
 
 		if ( is_wp_error( $statuses ) || empty( $statuses ) ) {
 			$statuses = [];
 		}
 
-		// Expand and order the statuses
-		$ordered_statuses = [];
-		$hold_to_end      = [];
-		foreach ( $statuses as $key => $status ) {
-			// Unencode and set all of our psuedo term meta because we need the position if it exists
-			$unencoded_description = $this->get_unencoded_description( $status->description );
-			if ( is_array( $unencoded_description ) ) {
-				foreach ( $unencoded_description as $key => $value ) {
-					$status->$key = $value;
-				}
-			}
-			// We require the position key later on (e.g. management table)
-			if ( ! isset( $status->position ) ) {
-				$status->position = false;
-			}
-			// Only add the status to the ordered array if it has a set position and doesn't conflict with another key
-			// Otherwise, hold it for later
-			if ( $status->position && ! array_key_exists( $status->position, $ordered_statuses ) ) {
-				$ordered_statuses[ (int) $status->position ] = $status;
-			} else {
-				$hold_to_end[] = $status;
-			}
-		}
-		// Sort the items numerically by key
-		ksort( $ordered_statuses, SORT_NUMERIC );
-		// Append all of the statuses that didn't have an existing position
-		foreach ( $hold_to_end as $unpositioned_status ) {
-			$ordered_statuses[] = $unpositioned_status;
-		}
+		// Add metadata to each term
+		$statuses = array_map( function ( $status ) {
+			$term_meta    = apply_filters( 'vw_register_custom_status_meta', [], $status );
+			$status->meta = $term_meta;
 
-		$ordered_statuses = array_values( $ordered_statuses );
+			return $status;
+		}, $statuses );
 
-		$this->custom_statuses_cache = $ordered_statuses;
-		return $ordered_statuses;
+		// Set the internal object cache
+		self::$custom_statuses_cache = $statuses;
+
+		return $statuses;
 	}
 
 	/**
 	 * Returns the a single status object based on ID, title, or slug
 	 *
-	 * @param string|int $string_or_int The status to search for, either by slug, name or ID
+	 * @param string $field The field to search by
+	 * @param int|string $value The value to search for
 	 * @return WP_Term|false $status The object for the matching status
 	 */
-	public function get_custom_status_by( $field, $value ) {
+	public static function get_custom_status_by( string $field, int|string $value ): WP_Term|false {
 		// We only support id, slug and name for lookup.
 		if ( ! in_array( $field, [ 'id', 'slug', 'name' ] ) ) {
 			return false;
 		}
 
+		$custom_status = false;
+
 		if ( 'id' === $field ) {
-			$field = 'term_id';
+			$custom_status = get_term( $value, self::TAXONOMY_KEY );
+		} else {
+			$custom_status = get_term_by( $field, $value, self::TAXONOMY_KEY );
 		}
 
-		// ToDo: This is inefficient as we are fetching all the terms, and then finding the one that matches.
-		$custom_statuses = $this->get_custom_statuses();
-		$custom_status   = wp_filter_object_list( $custom_statuses, [ $field => $value ] );
+		if ( is_wp_error( $custom_status ) || ! $custom_status ) {
+			$custom_status = false;
+		} else {
+			$term_meta           = apply_filters( 'vw_register_custom_status_meta', [], $custom_status );
+			$custom_status->meta = $term_meta;
+		}
 
-		// array_shift will ensure to set the first one or null if the array is empty
-		$custom_status = array_shift( $custom_status );
+		return $custom_status;
+	}
 
-		// If $custom_status is null, return false or else return the status object
-		return null !== $custom_status ? $custom_status : false;
+	/**
+	 * Get the core statuses that are used when the post type is unsupported
+	 *
+	 * Note: This is necessary because a new post does not have a post type available, and that causes a whole host of problems.
+	 *
+	 * @return array $default_terms The default statuses
+	 */
+	public static function get_core_statuses(): array {
+		$default_terms = [
+			[
+				'name'        => __( 'Draft', 'vip-workflow' ),
+				'slug'        => 'draft',
+				'description' => __( 'Post is a draft; not ready for review or publication.', 'vip-workflow' ),
+			],
+			[
+				'name'        => __( 'Pending Review' ),
+				'slug'        => 'pending',
+				'description' => __( 'Post needs to be reviewed by an editor.', 'vip-workflow' ),
+			],
+		];
+
+		return $default_terms;
 	}
 
 	/**
@@ -736,9 +853,9 @@ class Custom_Status extends Module {
 	 * @param string $new_status Slug for the new status
 	 * @return true|WP_Error
 	 */
-	public function reassign_post_status( $old_status, $new_status ) {
+	public static function reassign_post_status( string $old_status, string $new_status ): bool|WP_Error {
 		$old_status_post_ids = ( new WP_Query( [
-			'post_type'      => $this->get_supported_post_types(),
+			'post_type'      => HelperUtilities::get_supported_post_types(),
 			'post_status'    => $old_status,
 			'posts_per_page' => -1,
 			'fields'         => 'ids',
@@ -787,8 +904,8 @@ class Custom_Status extends Module {
 	 *
 	 * @return array $post_states
 	 */
-	public function add_status_to_post_states( $post_states, $post ) {
-		if ( ! in_array( $post->post_type, $this->get_supported_post_types(), true ) ) {
+	public static function add_status_to_post_states( array $post_states, WP_Post $post ): array {
+		if ( ! in_array( $post->post_type, HelperUtilities::get_supported_post_types(), true ) ) {
 			// Return early if this post type doesn't support custom statuses.
 			return $post_states;
 		}
@@ -819,9 +936,11 @@ class Custom_Status extends Module {
 	 * @param string $slug Slug of the status
 	 * @return bool $restricted True if restricted, false if not
 	 */
-	public function is_restricted_status( $slug ) {
+	public static function is_restricted_status( string $slug ): bool {
 
 		switch ( $slug ) {
+			case 'draft':
+			case 'pending':
 			case 'publish':
 			case 'private':
 			case 'future':
@@ -840,501 +959,24 @@ class Custom_Status extends Module {
 	}
 
 	/**
-	 * Primary configuration page for custom status class, which is also the main entry point for configuring the plugin
-	 */
-	public function print_configure_view() {
-		include_once __DIR__ . '/views/manage-workflow.php';
-	}
-
-	/**
-	 * Given a post ID, return true if the extended post status allows for publishing.
-	 *
-	 * @param int $post_id The post ID being queried.
-	 * @return bool True if the post should not be published based on the extended post status, false otherwise.
-	 */
-	public function workflow_is_publish_blocked( $post_id ) {
-		$post = get_post( $post_id );
-
-		if ( null === $post ) {
-			return false;
-		}
-
-		$custom_statuses = $this->get_custom_statuses();
-		$status_slugs    = wp_list_pluck( $custom_statuses, 'slug' );
-
-		if ( ! in_array( $post->post_status, $status_slugs ) || ! in_array( $post->post_type, $this->get_supported_post_types() ) ) {
-			// Post is not using a custom status, or is not a supported post type
-			return false;
-		}
-
-		$status_before_publish = $custom_statuses[ array_key_last( $custom_statuses ) ];
-
-		if ( $status_before_publish->slug === $post->post_status ) {
-			// Post is in the last status, so it can be published
-			return true;
-		} else {
-			return false;
-		}
-	}
-
-	/**
 	 * Given a post ID, return true if the post type is supported and using a custom status, false otherwise.
 	 *
 	 * @param int $post_id The post ID being queried.
 	 * @return bool True if the post is using a custom status, false otherwise.
 	 */
-	public function is_post_using_custom_status( $post_id ) {
+	public static function is_post_using_custom_status( int $post_id ): bool {
 		$post = get_post( $post_id );
 
 		if ( null === $post ) {
 			return false;
 		}
 
-		$custom_post_types = $this->get_supported_post_types();
-		$custom_statuses   = $this->get_custom_statuses();
+		$custom_post_types = HelperUtilities::get_supported_post_types();
+		$custom_statuses   = self::get_custom_statuses();
 		$status_slugs      = wp_list_pluck( $custom_statuses, 'slug' );
 
 		return in_array( $post->post_type, $custom_post_types ) && in_array( $post->post_status, $status_slugs );
 	}
-
-	/**
-	 * Register REST API endpoints for custom statuses
-	 */
-	public function register_rest_endpoints() {
-		EditStatus::init();
-	}
-
-	// Hacks for custom statuses to work with core
-
-	// phpcs:disable:WordPress.Security.NonceVerification.Missing -- Disabling nonce verification but we should renable it.
-
-	/**
-	 * This is a hack! hack! hack! until core is fixed/better supports custom statuses
-	 *
-	 * When publishing a post with a custom status, set the status to 'pending' temporarily
-	 * @see Works around this limitation: http://core.trac.wordpress.org/browser/tags/3.2.1/wp-includes/post.php#L2694
-	 * @see Original thread: http://wordpress.org/support/topic/plugin-edit-flow-custom-statuses-create-timestamp-problem
-	 * @see Core ticket: http://core.trac.wordpress.org/ticket/18362
-	 */
-	public function check_timestamp_on_publish() {
-		global $pagenow, $wpdb;
-
-		if ( $this->disable_custom_statuses_for_post_type() ) {
-			return;
-		}
-
-		// Handles the transition to 'publish' on edit.php
-		if ( VIP_Workflow::instance() !== null && 'edit.php' === $pagenow && isset( $_REQUEST['bulk_edit'] ) ) {
-			// For every post_id, set the post_status as 'pending' only when there's no timestamp set for $post_date_gmt
-			if ( isset( $_REQUEST['post'] ) && isset( $_REQUEST['_status'] ) && 'publish' === $_REQUEST['_status'] ) {
-				$post_ids = array_map( 'intval', (array) $_REQUEST['post'] );
-				foreach ( $post_ids as $post_id ) {
-					$wpdb->update( $wpdb->posts, [ 'post_status' => 'pending' ], [
-						'ID'            => $post_id,
-						'post_date_gmt' => '0000-00-00 00:00:00',
-					] );
-					clean_post_cache( $post_id );
-				}
-			}
-		}
-
-		// Handles the transition to 'publish' on post.php
-		if ( VIP_Workflow::instance() !== null && 'post.php' === $pagenow && isset( $_POST['publish'] ) ) {
-			// Set the post_status as 'pending' only when there's no timestamp set for $post_date_gmt
-			if ( isset( $_POST['post_ID'] ) ) {
-				$post_id = (int) $_POST['post_ID'];
-				$ret     = $wpdb->update( $wpdb->posts, [ 'post_status' => 'pending' ], [
-					'ID'            => $post_id,
-					'post_date_gmt' => '0000-00-00 00:00:00',
-				] );
-				clean_post_cache( $post_id );
-				foreach ( [ 'aa', 'mm', 'jj', 'hh', 'mn' ] as $timeunit ) {
-					if ( isset( $_POST[ $timeunit ] ) && ! empty( $_POST[ 'hidden_' . $timeunit ] ) && $_POST[ 'hidden_' . $timeunit ] != $_POST[ $timeunit ] ) {
-						$edit_date = '1';
-						break;
-					}
-				}
-				if ( $ret && empty( $edit_date ) ) {
-					add_filter( 'pre_post_date', [ $this, 'helper_timestamp_hack' ] );
-					add_filter( 'pre_post_date_gmt', [ $this, 'helper_timestamp_hack' ] );
-				}
-			}
-		}
-	}
-
-	//phpcs:enable:WordPress.Security.NonceVerification.Missing
-
-	/**
-	 * PHP < 5.3.x doesn't support anonymous functions
-	 * This helper is only used for the check_timestamp_on_publish method above
-	 */
-	public function helper_timestamp_hack() {
-		return ( 'pre_post_date' === current_filter() ) ? current_time( 'mysql' ) : '';
-	}
-
-	/**
-	 * This is a hack! hack! hack! until core is fixed/better supports custom statuses
-	 *
-	 * Normalize post_date_gmt if it isn't set to the past or the future
-	 * @see Works around this limitation: https://core.trac.wordpress.org/browser/tags/4.5.1/src/wp-includes/post.php#L3182
-	 * @see Original thread: http://wordpress.org/support/topic/plugin-edit-flow-custom-statuses-create-timestamp-problem
-	 * @see Core ticket: http://core.trac.wordpress.org/ticket/18362
-	 */
-	public function fix_custom_status_timestamp( $data, $postarr ) {
-		// Don't run this if VIP Workflow isn't active, or we're on some other page
-		if ( $this->disable_custom_statuses_for_post_type()
-		|| VIP_Workflow::instance() === null ) {
-			return $data;
-		}
-
-		$status_slugs = wp_list_pluck( $this->get_custom_statuses(), 'slug' );
-
-		//Post is scheduled or published? Ignoring.
-		if ( ! in_array( $postarr['post_status'], $status_slugs ) ) {
-			return $data;
-		}
-
-		//If empty, keep empty.
-		if ( empty( $postarr['post_date_gmt'] )
-		|| '0000-00-00 00:00:00' === $postarr['post_date_gmt'] ) {
-			$data['post_date_gmt'] = '0000-00-00 00:00:00';
-		}
-
-		return $data;
-	}
-
-	/**
-	 * A new hack! hack! hack! until core better supports custom statuses`
-	 *
-	 * If the post_name is set, set it, otherwise keep it empty
-	 *
-	 * @see https://github.com/Automattic/Edit-Flow/issues/523
-	 * @see https://github.com/Automattic/Edit-Flow/issues/633
-	 */
-	public function maybe_keep_post_name_empty( $data, $postarr ) {
-		$status_slugs = wp_list_pluck( $this->get_custom_statuses(), 'slug' );
-
-		// Ignore if it's not a post status and post type we support
-		if ( ! in_array( $data['post_status'], $status_slugs )
-		|| ! in_array( $data['post_type'], $this->get_supported_post_types() ) ) {
-			return $data;
-		}
-
-		// If the post_name was intentionally set, set the post_name
-		if ( ! empty( $postarr['post_name'] ) ) {
-			$data['post_name'] = sanitize_title( $postarr['post_name'] );
-			return $data;
-		}
-
-		// Otherwise, keep the post_name empty
-		$data['post_name'] = '';
-
-		return $data;
-	}
-
-	/**
-	 * A new hack! hack! hack! until core better supports custom statuses`
-	 *
-	 * `wp_unique_post_slug` is used to set the `post_name`. When a custom status is used, WordPress will try
-	 * really hard to set `post_name`, and we leverage `wp_unique_post_slug` to prevent it being set
-	 *
-	 * @see: https://github.com/WordPress/WordPress/blob/396647666faebb109d9cd4aada7bb0c7d0fb8aca/wp-includes/post.php#L3932
-	*/
-	public function fix_unique_post_slug( $override_slug, $slug, $post_ID, $post_status, $post_type, $post_parent ) {
-		$status_slugs = wp_list_pluck( $this->get_custom_statuses(), 'slug' );
-
-		if ( ! in_array( $post_status, $status_slugs )
-		|| ! in_array( $post_type, $this->get_supported_post_types() ) ) {
-			return null;
-		}
-
-		$post = get_post( $post_ID );
-
-		if ( empty( $post ) ) {
-			return null;
-		}
-
-		if ( $post->post_name ) {
-			return $slug;
-		}
-
-		return '';
-	}
-
-
-	/**
-	 * Another hack! hack! hack! until core better supports custom statuses
-	 *
-	 * The preview link for an unpublished post should always be ?p=
-	 */
-	public function fix_preview_link_part_one( $preview_link ) {
-		global $pagenow;
-
-		$post = get_post( get_the_ID() );
-
-		// Only modify if we're using a pre-publish status on a supported custom post type
-		$status_slugs = wp_list_pluck( $this->get_custom_statuses(), 'slug' );
-		if ( ! $post
-		|| ! is_admin()
-		|| 'post.php' != $pagenow
-		|| ! in_array( $post->post_status, $status_slugs )
-		|| ! in_array( $post->post_type, $this->get_supported_post_types() )
-		|| strpos( $preview_link, 'preview_id' ) !== false
-		|| 'sample' === $post->filter ) {
-			return $preview_link;
-		}
-
-		return $this->get_preview_link( $post );
-	}
-
-	/**
-	 * Another hack! hack! hack! until core better supports custom statuses
-	 *
-	 * The preview link for an unpublished post should always be ?p=
-	 * The code used to trigger a post preview doesn't also apply the 'preview_post_link' filter
-	 * So we can't do a targeted filter. Instead, we can even more hackily filter get_permalink
-	 * @see http://core.trac.wordpress.org/ticket/19378
-	 */
-	public function fix_preview_link_part_two( $permalink, $post, $sample ) {
-		global $pagenow;
-
-		if ( is_int( $post ) ) {
-			$post = get_post( $post );
-		}
-
-		//Should we be doing anything at all?
-		if ( ! in_array( $post->post_type, $this->get_supported_post_types() ) ) {
-			return $permalink;
-		}
-
-		//Is this published?
-		if ( in_array( $post->post_status, $this->published_statuses ) ) {
-			return $permalink;
-		}
-
-		//Are we overriding the permalink? Don't do anything
-		// phpcs:ignore:WordPress.Security.NonceVerification.Missing
-		if ( isset( $_POST['action'] ) && 'sample-permalink' === $_POST['action'] ) {
-			return $permalink;
-		}
-
-		//Are we previewing the post from the normal post screen?
-		if ( ( 'post.php' === $pagenow || 'post-new.php' === $pagenow )
-		// phpcs:ignore:WordPress.Security.NonceVerification.Missing
-		&& ! isset( $_POST['wp-preview'] ) ) {
-			return $permalink;
-		}
-
-		//If it's a sample permalink, not a preview
-		if ( $sample ) {
-			return $permalink;
-		}
-
-		return $this->get_preview_link( $post );
-	}
-
-	/**
-	 * Another hack! hack! hack! until core better supports custom statuses
-	 *
-	 * The preview link for a saved unpublished post with a custom status returns a 'preview_nonce'
-	 * in it and needs to be removed when previewing it to return a viewable preview link.
-	 * @see https://github.com/Automattic/Edit-Flow/issues/513
-	 */
-	public function fix_preview_link_part_three( $preview_link, $query_args ) {
-		$autosave = wp_get_post_autosave( $query_args->ID, get_current_user_id() );
-		if ( $autosave ) {
-			foreach ( array_intersect( array_keys( _wp_post_revision_fields( $query_args ) ), array_keys( _wp_post_revision_fields( $autosave ) ) ) as $field ) {
-				if ( normalize_whitespace( $query_args->$field ) != normalize_whitespace( $autosave->$field ) ) {
-					// Pass through, it's a personal preview.
-					return $preview_link;
-				}
-			}
-		}
-		return remove_query_arg( [ 'preview_nonce' ], $preview_link );
-	}
-
-	/**
-	 * Fix get_sample_permalink. Previosuly the 'editable_slug' filter was leveraged
-	 * to correct the sample permalink a user could edit on post.php. Since 4.4.40
-	 * the `get_sample_permalink` filter was added which allows greater flexibility in
-	 * manipulating the slug. Critical for cases like editing the sample permalink on
-	 * hierarchical post types.
-	 *
-	 * @param string  $permalink Sample permalink
-	 * @param int     $post_id   Post ID
-	 * @param string  $title     Post title
-	 * @param string  $name      Post name (slug)
-	 * @param WP_Post $post      Post object
-	 * @return string $link Direct link to complete the action
-	 */
-	public function fix_get_sample_permalink( $permalink, $post_id, $title, $name, $post ) {
-
-		$status_slugs = wp_list_pluck( $this->get_custom_statuses(), 'slug' );
-
-		if ( ! in_array( $post->post_status, $status_slugs )
-		|| ! in_array( $post->post_type, $this->get_supported_post_types() ) ) {
-			return $permalink;
-		}
-
-		remove_filter( 'get_sample_permalink', [ $this, 'fix_get_sample_permalink' ], 10, 5 );
-
-		$new_name  = ! is_null( $name ) ? $name : $post->post_name;
-		$new_title = ! is_null( $title ) ? $title : $post->post_title;
-
-		$post              = get_post( $post_id );
-		$status_before     = $post->post_status;
-		$post->post_status = 'draft';
-
-		$permalink = get_sample_permalink( $post, $title, sanitize_title( $new_name ? $new_name : $new_title, $post->ID ) );
-
-		$post->post_status = $status_before;
-
-		add_filter( 'get_sample_permalink', [ $this, 'fix_get_sample_permalink' ], 10, 5 );
-
-		return $permalink;
-	}
-
-	/**
-	 * Hack to work around post status check in get_sample_permalink_html
-	 *
-	 *
-	 * The get_sample_permalink_html checks the status of the post and if it's
-	 * a draft generates a certain permalink structure.
-	 * We need to do the same work it's doing for custom statuses in order
-	 * to support this link
-	 * @see https://core.trac.wordpress.org/browser/tags/4.5.2/src/wp-admin/includes/post.php#L1296
-	 *
-	 * @param string  $return    Sample permalink HTML markup.
-	 * @param int     $post_id   Post ID.
-	 * @param string  $new_title New sample permalink title.
-	 * @param string  $new_slug  New sample permalink slug.
-	 * @param WP_Post $post      Post object.
-	 */
-	public function fix_get_sample_permalink_html( $permalink, $post_id, $new_title, $new_slug, $post ) {
-		$status_slugs = wp_list_pluck( $this->get_custom_statuses(), 'slug' );
-
-		if ( ! in_array( $post->post_status, $status_slugs )
-		|| ! in_array( $post->post_type, $this->get_supported_post_types() ) ) {
-			return $permalink;
-		}
-
-		remove_filter( 'get_sample_permalink_html', [ $this, 'fix_get_sample_permalink_html' ], 10, 5 );
-
-		$post->post_status     = 'draft';
-		$sample_permalink_html = get_sample_permalink_html( $post, $new_title, $new_slug );
-
-		add_filter( 'get_sample_permalink_html', [ $this, 'fix_get_sample_permalink_html' ], 10, 5 );
-
-		return $sample_permalink_html;
-	}
-
-
-	/**
-	 * Fixes a bug where post-pagination doesn't work when previewing a post with a custom status
-	 * @link https://github.com/Automattic/Edit-Flow/issues/192
-	 *
-	 * This filter only modifies output if `is_preview()` is true
-	 *
-	 * Used by `wp_link_pages_link` filter
-	 *
-	 * @param $link
-	 * @param $i
-	 *
-	 * @return string
-	 */
-	public function modify_preview_link_pagination_url( $link, $i ) {
-
-		// Use the original $link when not in preview mode
-		if ( ! is_preview() ) {
-			return $link;
-		}
-
-		// Get an array of valid custom status slugs
-		$custom_statuses = wp_list_pluck( $this->get_custom_statuses(), 'slug' );
-
-		// Apply original link filters from core `wp_link_pages()`
-		$r = apply_filters( 'wp_link_pages_args', [
-			'link_before' => '',
-			'link_after'  => '',
-			'pagelink'    => '%',
-		]);
-
-		// _wp_link_page() && _vw_wp_link_page() produce an opening link tag ( <a href=".."> )
-		// This is necessary to replicate core behavior:
-		$link = $r['link_before'] . str_replace( '%', $i, $r['pagelink'] ) . $r['link_after'];
-		$link = _vw_wp_link_page( $i, $custom_statuses ) . $link . '</a>';
-
-
-		return $link;
-	}
-
-	/**
-	 * Get the proper preview link for a post
-	 */
-	private function get_preview_link( $post ) {
-
-		if ( 'page' === $post->post_type ) {
-			$args = [
-				'page_id' => $post->ID,
-			];
-		} elseif ( 'post' === $post->post_type ) {
-			$args = [
-				'p'       => $post->ID,
-				'preview' => 'true',
-			];
-		} else {
-			$args = [
-				'p'         => $post->ID,
-				'post_type' => $post->post_type,
-			];
-		}
-
-		$args['preview_id'] = $post->ID;
-		return add_query_arg( $args, home_url( '/' ) );
-	}
-
-	/**
-	 * Another hack! hack! hack! until core better supports custom statuses
-	 *
-	 * The preview link for an unpublished post should always be ?p=, even in the list table
-	 * @see http://core.trac.wordpress.org/ticket/19378
-	 */
-	public function fix_post_row_actions( $actions, $post ) {
-		global $pagenow;
-
-		// Only modify if we're using a pre-publish status on a supported custom post type
-		$status_slugs = wp_list_pluck( $this->get_custom_statuses(), 'slug' );
-		if ( 'edit.php' != $pagenow
-		|| ! in_array( $post->post_status, $status_slugs )
-		|| ! in_array( $post->post_type, $this->get_supported_post_types() ) ) {
-			return $actions;
-		}
-
-		// 'view' is only set if the user has permission to post
-		if ( empty( $actions['view'] ) ) {
-			return $actions;
-		}
-
-		if ( 'page' === $post->post_type ) {
-			$args = [
-				'page_id' => $post->ID,
-			];
-		} elseif ( 'post' === $post->post_type ) {
-			$args = [
-				'p' => $post->ID,
-			];
-		} else {
-			$args = [
-				'p'         => $post->ID,
-				'post_type' => $post->post_type,
-			];
-		}
-		$args['preview'] = 'true';
-		$preview_link    = add_query_arg( $args, home_url( '/' ) );
-
-		/* translators: %s: post title */
-		$actions['view'] = '<a href="' . esc_url( $preview_link ) . '" title="' . esc_attr( sprintf( __( 'Preview &#8220;%s&#8221;' ), $post->post_title ) ) . '" rel="permalink">' . __( 'Preview' ) . '</a>';
-		return $actions;
-	}
 }
 
+CustomStatus::init();
