@@ -1,20 +1,20 @@
 <?php
 /**
- * class EditStatus
- * REST endpoint for updating a custom status
+ * class CustomStatusEndpoint
+ * REST endpoint for managing custom statuses
  */
 
 namespace VIPWorkflow\Modules\CustomStatus\REST;
 
-use VIPWorkflow\Modules\Custom_Status;
-use VIPWorkflow\VIP_Workflow;
+use VIPWorkflow\Modules\CustomStatus;
+use VIPWorkflow\Modules\EditorialMetadata;
 use WP_Error;
 use WP_REST_Request;
 use WP_Term;
 
 defined( 'ABSPATH' ) || exit;
 
-class EditStatus {
+class CustomStatusEndpoint {
 	/**
 	 * Initialize the class
 	 */
@@ -32,7 +32,7 @@ class EditStatus {
 			'permission_callback' => [ __CLASS__, 'permission_callback' ],
 			'args'                => [
 				// Required parameters
-				'name'        => [
+				'name'              => [
 					'required'          => true,
 					'validate_callback' => function ( $param ) {
 						return ! empty( trim( $param ) );
@@ -43,10 +43,30 @@ class EditStatus {
 				],
 
 				// Optional parameters
-				'description' => [
+				'description'       => [
 					'default'           => '',
 					'sanitize_callback' => function ( $param ) {
 						return stripslashes( wp_filter_nohtml_kses( trim( $param ) ) );
+					},
+				],
+				'required_metadata_ids' => [
+					'default'           => [],
+					'validate_callback' => function ( $param ) {
+						$metadata_ids = array_map( 'absint', $param );
+						return self::is_valid_editorial_metadata_ids( $metadata_ids );
+					},
+					'sanitize_callback' => function ( $param ) {
+						return array_map( 'absint', $param );
+					},
+				],
+				'required_user_ids' => [
+					'default'           => [],
+					'validate_callback' => function ( $param ) {
+						$user_ids = array_map( 'absint', $param );
+						return self::is_valid_user_ids( $user_ids );
+					},
+					'sanitize_callback' => function ( $param ) {
+						return array_map( 'absint', $param );
 					},
 				],
 			],
@@ -58,7 +78,7 @@ class EditStatus {
 			'permission_callback' => [ __CLASS__, 'permission_callback' ],
 			'args'                => [
 				// Required parameters
-				'name'        => [
+				'name'              => [
 					'required'          => true,
 					'validate_callback' => function ( $param ) {
 						return ! empty( trim( $param ) );
@@ -67,11 +87,11 @@ class EditStatus {
 						return trim( $param );
 					},
 				],
-				'id'          => [
+				'id'                => [
 					'required'          => true,
 					'validate_callback' => function ( $param ) {
 						$term_id = absint( $param );
-						$term    = get_term( $term_id, Custom_Status::TAXONOMY_KEY );
+						$term    = get_term( $term_id, CustomStatus::TAXONOMY_KEY );
 						return ( $term instanceof WP_Term );
 					},
 					'sanitize_callback' => function ( $param ) {
@@ -80,10 +100,30 @@ class EditStatus {
 				],
 
 				// Optional parameters
-				'description' => [
+				'description'       => [
 					'default'           => '',
 					'sanitize_callback' => function ( $param ) {
 						return stripslashes( wp_filter_nohtml_kses( trim( $param ) ) );
+					},
+				],
+				'required_metadata_ids' => [
+					'default'           => [],
+					'validate_callback' => function ( $param ) {
+						$metadata_ids = array_map( 'absint', $param );
+						return self::is_valid_editorial_metadata_ids( $metadata_ids );
+					},
+					'sanitize_callback' => function ( $param ) {
+						return array_map( 'absint', $param );
+					},
+				],
+				'required_user_ids' => [
+					'default'           => [],
+					'validate_callback' => function ( $param ) {
+						$user_ids = array_map( 'absint', $param );
+						return self::is_valid_user_ids( $user_ids );
+					},
+					'sanitize_callback' => function ( $param ) {
+						return array_map( 'absint', $param );
 					},
 				],
 			],
@@ -99,7 +139,7 @@ class EditStatus {
 					'required'          => true,
 					'validate_callback' => function ( $param ) {
 						$term_id = absint( $param );
-						$term    = get_term( $term_id, Custom_Status::TAXONOMY_KEY );
+						$term    = get_term( $term_id, CustomStatus::TAXONOMY_KEY );
 						return ( $term instanceof WP_Term );
 					},
 					'sanitize_callback' => function ( $param ) {
@@ -125,7 +165,7 @@ class EditStatus {
 						// validate each item in the array.
 						foreach ( $param as $position => $term_id ) {
 							$term_id = absint( $term_id );
-							$term    = get_term( $term_id, Custom_Status::TAXONOMY_KEY );
+							$term    = get_term( $term_id, CustomStatus::TAXONOMY_KEY );
 							if ( ! $term instanceof WP_Term ) {
 								return false;
 							}
@@ -158,11 +198,11 @@ class EditStatus {
 	 * @param WP_REST_Request $request
 	 */
 	public static function handle_create_status( WP_REST_Request $request ) {
-		$status_name        = sanitize_text_field( $request->get_param( 'name' ) );
-		$status_slug        = sanitize_title( $request->get_param( 'name' ) );
-		$status_description = $request->get_param( 'description' );
-
-		$custom_status_module = VIP_Workflow::instance()->custom_status;
+		$status_name               = sanitize_text_field( $request->get_param( 'name' ) );
+		$status_slug               = sanitize_title( $request->get_param( 'name' ) );
+		$status_description        = $request->get_param( 'description' );
+		$status_required_metadata_ids = $request->get_param( 'required_metadata_ids' );
+		$status_required_user_ids = $request->get_param( 'required_user_ids' );
 
 		// Check that the name isn't numeric
 		if ( is_numeric( $status_name ) ) {
@@ -170,7 +210,7 @@ class EditStatus {
 		}
 
 		// Check to make sure the name is not restricted
-		if ( $custom_status_module->is_restricted_status( strtolower( $status_name ) ) ) {
+		if ( CustomStatus::is_restricted_status( strtolower( $status_name ) ) ) {
 			return new WP_Error( 'invalid', 'Status name is restricted. Please chose another name.' );
 		}
 
@@ -180,19 +220,21 @@ class EditStatus {
 		}
 
 		// Check to make sure the status doesn't already exist as another term because otherwise we'd get a fatal error
-		$term_exists = term_exists( $status_slug, Custom_Status::TAXONOMY_KEY );
+		$term_exists = term_exists( $status_slug, CustomStatus::TAXONOMY_KEY );
 
 		if ( $term_exists ) {
 			return new WP_Error( 'invalid', 'Status name conflicts with existing term. Please choose another.' );
 		}
 
-		// get status_slug & status_description
 		$args = [
-			'description' => $status_description,
-			'slug'        => $status_slug,
+			'name'               => $status_name,
+			'description'        => $status_description,
+			'slug'               => $status_slug,
+			'required_metadata_ids' => $status_required_metadata_ids,
+			'required_user_ids' => $status_required_user_ids,
 		];
 
-		$add_status_result = $custom_status_module->add_custom_status( $status_name, $args );
+		$add_status_result = CustomStatus::add_custom_status( $args );
 
 		// Regardless of an error being thrown, the result will be returned so the client can handle it.
 		return rest_ensure_response( $add_status_result );
@@ -204,12 +246,12 @@ class EditStatus {
 	 * @param WP_REST_Request $request
 	 */
 	public static function handle_update_status( WP_REST_Request $request ) {
-		$term_id            = $request->get_param( 'id' );
-		$status_name        = sanitize_text_field( $request->get_param( 'name' ) );
-		$status_slug        = sanitize_title( $request->get_param( 'name' ) );
-		$status_description = $request->get_param( 'description' );
-
-		$custom_status_module = VIP_Workflow::instance()->custom_status;
+		$term_id                   = $request->get_param( 'id' );
+		$status_name               = sanitize_text_field( $request->get_param( 'name' ) );
+		$status_slug               = sanitize_title( $request->get_param( 'name' ) );
+		$status_description        = $request->get_param( 'description' );
+		$status_required_metadata_ids = $request->get_param( 'required_metadata_ids' );
+		$status_required_user_ids = $request->get_param( 'required_user_ids' );
 
 		// Check that the name isn't numeric
 		if ( is_numeric( $status_name ) ) {
@@ -217,7 +259,7 @@ class EditStatus {
 		}
 
 		// Check to make sure the name is not restricted
-		if ( $custom_status_module->is_restricted_status( strtolower( $status_name ) ) ) {
+		if ( CustomStatus::is_restricted_status( strtolower( $status_name ) ) ) {
 			return new WP_Error( 'invalid', 'Status name is restricted. Please chose another name.' );
 		}
 
@@ -227,16 +269,15 @@ class EditStatus {
 		}
 
 		// Check to make sure the status doesn't already exist
-		$custom_status_by_id = $custom_status_module->get_custom_status_by( 'id', $term_id );
-
-		$custom_status_by_slug = $custom_status_module->get_custom_status_by( 'slug', $status_slug );
+		$custom_status_by_id   = CustomStatus::get_custom_status_by( 'id', $term_id );
+		$custom_status_by_slug = CustomStatus::get_custom_status_by( 'slug', $status_slug );
 
 		if ( $custom_status_by_slug && $custom_status_by_id && $custom_status_by_id->slug !== $status_slug ) {
 			return new WP_Error( 'invalid', 'Status already exists. Please choose another name.' );
 		}
 
 		// Check to make sure the status doesn't already exist as another term because otherwise we'd get a fatal error
-		$term_exists = term_exists( $status_slug, Custom_Status::TAXONOMY_KEY );
+		$term_exists = term_exists( $status_slug, CustomStatus::TAXONOMY_KEY );
 
 		// term_id from term_exists is a string, while term_id is an integer so not using strict comparison
 		if ( $term_exists && isset( $term_exists['term_id'] ) && $term_exists['term_id'] != $term_id ) {
@@ -245,16 +286,17 @@ class EditStatus {
 
 		// get status_name & status_description
 		$args = [
-			'name'        => $status_name,
-			'description' => $status_description,
-			'slug'        => $status_slug,
+			'name'               => $status_name,
+			'description'        => $status_description,
+			'slug'               => $status_slug,
+			'required_metadata_ids' => $status_required_metadata_ids,
+			'required_user_ids' => $status_required_user_ids,
 		];
 
-		// ToDo: Ensure that we don't do an update when the name and description are the same as the current status
-		$update_status_result = $custom_status_module->update_custom_status( $term_id, $args );
+		$updated_status = CustomStatus::update_custom_status( $term_id, $args );
 
 		// Regardless of an error being thrown, the result will be returned so the client can handle it.
-		return rest_ensure_response( $update_status_result );
+		return rest_ensure_response( $updated_status );
 	}
 
 	/**
@@ -265,15 +307,13 @@ class EditStatus {
 	public static function handle_delete_status( WP_REST_Request $request ) {
 		$term_id = $request->get_param( 'id' );
 
-		$custom_status_module = VIP_Workflow::instance()->custom_status;
-
 		// Check to make sure the status exists
-		$custom_status_by_id = $custom_status_module->get_custom_status_by( 'id', $term_id );
+		$custom_status_by_id = CustomStatus::get_custom_status_by( 'id', $term_id );
 		if ( ! $custom_status_by_id ) {
 			return new WP_Error( 'invalid', 'Status does not exist.' );
 		}
 
-		$delete_status_result = $custom_status_module->delete_custom_status( $term_id );
+		$delete_status_result = CustomStatus::delete_custom_status( $term_id );
 
 		// Regardless of an error being thrown, the result will be returned so the client can handle it.
 		return rest_ensure_response( $delete_status_result );
@@ -287,13 +327,10 @@ class EditStatus {
 	public static function handle_reorder_status( WP_REST_Request $request ) {
 		$status_order = $request->get_param( 'status_positions' );
 
-		$custom_status_module = VIP_Workflow::instance()->custom_status;
-
 		if ( ! is_array( $status_order ) ) {
 			return new WP_Error( 'invalid', 'Status order must be an array.' );
 		}
 
-		// ToDo: Switch this to be a bulk update instead.
 		foreach ( $status_order as $position => $term_id ) {
 
 			// Have to add 1 to the position because the index started with zero
@@ -301,7 +338,7 @@ class EditStatus {
 				'position' => absint( $position ) + 1,
 			];
 
-			$update_status_result = $custom_status_module->update_custom_status( (int) $term_id, $args );
+			$update_status_result = CustomStatus::update_custom_status( (int) $term_id, $args );
 
 			// Stop the operation immediately if something has gone wrong, rather than silently continuing.
 			if ( is_wp_error( $update_status_result ) ) {
@@ -310,7 +347,45 @@ class EditStatus {
 		}
 
 		// Regardless of an error being thrown, the result will be returned so the client can handle it.
-		return rest_ensure_response( $custom_status_module->get_custom_statuses() );
+		return rest_ensure_response( CustomStatus::get_custom_statuses() );
+	}
+
+	// Utility functions
+
+	private static function is_valid_editorial_metadata_ids( $metadata_ids ) {
+		// If the array is empty, it's valid
+		if ( [] === $metadata_ids ) {
+			return true;
+		}
+
+		foreach ( $metadata_ids as $metadata_id ) {
+			$editorial_metadata = EditorialMetadata::get_editorial_metadata_term_by( 'id', $metadata_id );
+			if ( is_wp_error( $editorial_metadata ) || ! $editorial_metadata ) {
+				return false;
+			}
+		}
+
+		return true;
+	}
+
+	private static function is_valid_user_ids( $user_ids ) {
+		// If the array is empty, it's valid
+		if ( [] === $user_ids ) {
+			return true;
+		}
+
+		$args = [
+			'include' => $user_ids,
+			'fields'  => [ 'ID' ],
+		];
+
+		$users_found = get_users( $args );
+
+		if ( count( $users_found ) !== count( $user_ids ) ) {
+			return false;
+		}
+
+		return true;
 	}
 
 	// Public API
@@ -333,3 +408,5 @@ class EditStatus {
 		return rest_url( sprintf( '%s/%s', VIP_WORKFLOW_REST_NAMESPACE, 'custom-status/reorder' ) );
 	}
 }
+
+CustomStatusEndpoint::init();
