@@ -4,4 +4,11 @@ defined( 'ABSPATH' ) || exit();
 
 ?>
 
-<div id="editorial-metadata-manager"></div>
+<div class="wrap vip-workflow-admin">
+	<div class="explanation">
+		<h3><?php esc_html_e( 'Configure your editorial metadata for each post.', 'vip-workflow' ); ?></h3>
+		<p><?php esc_html_e( 'Editorial metadata can be used to enhance the details logged with each post, and is completely customizable for your workflow using a variety of supported field types.', 'vip-workflow' ); ?></p>
+	</div>
+
+	<div id="editorial-metadata-manager"></div>
+</div>
