@@ -2,7 +2,7 @@ const defaultScriptsConfig = require( '@wordpress/scripts/config/webpack.config'
 const glob = require( 'glob' );
 require( 'dotenv' ).config();
 
-const wpScriptsModules = [ 'custom-status', 'editorial-metadata', 'preview' ];
+const wpScriptsModules = [ 'custom-status', 'editorial-metadata', 'preview', 'settings' ];
 const wpScriptsModulesGlob =
 	wpScriptsModules.length === 1 ? wpScriptsModules[ 0 ] : `{${ wpScriptsModules.join( ',' ) }}`;
 
@@ -21,6 +21,23 @@ module.exports = [
 		...defaultScriptsConfig,
 		entry: {
 			...wpScriptsModulesEntries,
+		},
+		module: {
+			...defaultScriptsConfig.module,
+			rules: [
+				...defaultScriptsConfig.module.rules,
+				{
+					test: /\.tsx?$/,
+					use: [
+						{
+							loader: 'ts-loader',
+							options: {
+								transpileOnly: true,
+							},
+						},
+					],
+				},
+			],
 		},
 		output: {
 			...defaultScriptsConfig.output,
